feature: the choice-first rebuild becomes THE overworld builder
The parallel rebuild replaces the old pipeline wholesale and drops its
provisional name — there is no v2, only overworld_build:

- New build() entry: per-seed budget allotment, then per world the
  knob-free phases (Connectivity -> Levels -> Forts -> Locks) plus the
  diagnosis-driven Shaping loop, wrapped in pipe-web redeals; then the
  cross-world passes (secret-exit-safety backstop, hammer-bro fill,
  toad-house/spade promotion, sprite redistribution) reusing the shared
  capacity machinery verbatim.
- New HammerBroFill phase: every leftover reachable blank becomes a
  HammerBro slot (BFS-nearest within the pointer-table budget), with
  mandatory slots pinned under vanilla wandering-sprite positions when the
  hammer-bro shuffle is off — the same contract the old builder honored.
- Fort sections renumbered dense 0..n before the writer (it pairs locks to
  forts by index; the Locks phase may remove an unlockable fort).
- Old pipeline deleted: sections/shape/knobs/scoring/pipes/locks/
  gated_shortcut and their ~24 internals-bound tests. Shared machinery
  (types, capacity, route_choice, progression) stays; bfs_ordered and the
  shared placement helpers moved into capacity.rs. LockAssignment loses
  the now-unread blocks_target field. No builder-related CLI flags existed
  to remove.
- Census/env surface renamed: test_v2_* -> test_builder_*, V2_SEEDS ->
  CENSUS_SEEDS.

Verification:
- cargo clippy --all-targets: zero warnings; full suite green in debug
  (writer debug_asserts active) and release; wasm-pack build OK.
- test_full_determinism (byte-identical ROMs per seed) passes; CLI smoke
  ROMs byte-identical across runs and generate under heavy flag mixes.
- New test_builder_output_completable: real build() output passes the
  order-free completability fixpoint, one-lock-per-fort, dense sections,
  honest secret-exit flags, per-seed cross-world safety.
- 1000-seed shaping census identical column-for-column to the pre-move
  run (C1min 14 / sub-floor 0% every world; linear W7 58 / W4 15 /
  overall ~16%; 27.4 ms/seed native).
- ROM-level required_progression.py sane on generated ROMs (its W8
  eights-are-wild UNREACHABLE is a known analyzer gap: it models only W3
  canoes).
