--- src/randomize/overworld_build/knobs.rs
@@ -1,258 +1,0 @@
-//! Scoring knobs for every placement pass, bundled in one [`Knobs`] struct.
-//!
-//! This is the single place to look for the numeric personality of the
-//! builder. Placement passes each read ONLY their own section (`knobs.fort`,
-//! `knobs.level`, `knobs.lock`, `knobs.pipe`).
-//!
-//! Progression shape is NOT decided here: the choice-first builder measures
-//! the built route structure (`route_choice`) and shapes forts/locks/pipes
-//! against it, so there is no sampled archetype. Knobs are soft biases and
-//! can never fail — geometry that resists a bias degrades the aesthetics,
-//! never completability.
-
-/// Whether the spare-pipe pass may realize a fort-skip (a pipe bridging
-/// across one closed lock, making that fort optional).
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) enum FortSkipPolicy {
-    /// No pipe may bypass any mandatory fort — pairs that would are rejected.
-    // Reason: not yet used — every world currently allows AnyOneFort. This
-    // variant is the "no skip" arm for a future per-world roll; the consuming
-    // check in place_spare_pipes already honors it.
-    #[allow(dead_code)]
-    Never,
-    /// The first candidate pair that bypasses exactly ONE mandatory fort is
-    /// taken (the prize), at most once per world. Pairs bypassing 2+ forts,
-    /// or opening the goal with every lock closed, are always rejected
-    /// regardless of policy.
-    AnyOneFort,
-}
-
-/// Spread/density weights shared by the level and fortress scorers
-/// (`score_with_weights`). Embedded separately in [`FortScoring`] and
-/// [`LevelScoring`] so the two can diverge independently.
-#[derive(Clone, Copy, Debug)]
-pub(crate) struct SpreadScoring {
-    /// Weight on min Manhattan distance to already-placed content
-    /// (visual/spatial spread).
-    pub manhattan: f64,
-    /// Weight on min BFS-distance difference (traversal spread — weighted
-    /// above grid distance so spread follows the walk, not the screen).
-    pub bfs: f64,
-    /// Penalty per already-placed tile within `density_radius`.
-    pub density_penalty: f64,
-    /// Combined manhattan+BFS distance that counts as "nearby" for the
-    /// density penalty.
-    pub density_radius: usize,
-    /// Cap on each separation metric's contribution, so one huge gap can't
-    /// dominate the score.
-    pub separation_cap: f64,
-}
-
-impl Default for SpreadScoring {
-    fn default() -> Self {
-        SpreadScoring {
-            manhattan: 1.0,
-            bfs: 1.5,
-            density_penalty: 3.0,
-            density_radius: 4,
-            separation_cap: 8.0,
-        }
-    }
-}
-
-/// Knobs for fortress placement (the `shape_forts` pass — candidate ranking
-/// in the measured phase, softmax sampling in the aesthetic phase).
-#[derive(Clone, Copy, Debug)]
-pub(crate) struct FortScoring {
-    pub spread: SpreadScoring,
-    /// Bonus for one-exit positions — fortresses naturally belong at path
-    /// termini.
-    pub dead_end_bonus: f64,
-    /// Bonus for the designated island positions in
-    /// `FORTRESS_BONUS_POSITIONS` (isolated spots that rarely win placement
-    /// without a boost).
-    pub island_bonus: f64,
-    /// Softmax temperature. Score range ≈ [-12, +15] including the dead-end
-    /// bonus; 4.0 keeps top candidates favored without determinism.
-    pub softmax_t: f64,
-}
-
-impl Default for FortScoring {
-    fn default() -> Self {
-        FortScoring {
-            spread: SpreadScoring::default(),
-            dead_end_bonus: 5.0,
-            island_bonus: 0.5,
-            softmax_t: 4.0,
-        }
-    }
-}
-
-/// Knobs for level placement (`place_levels`).
-#[derive(Clone, Copy, Debug)]
-pub(crate) struct LevelScoring {
-    pub spread: SpreadScoring,
-    /// Bonus for one-exit positions (small — levels only mildly prefer
-    /// dead ends).
-    pub dead_end_bonus: f64,
-    /// Path-relevance weight: positions on the main start→target route (low
-    /// detour) score higher. Max bonus = `path_detour_cap * path_bonus`.
-    ///
-    /// THE dominant linearity lever — and a deliberate choice↔streak trade.
-    /// Raising it puts more levels on the trunk, which the route scorer reads
-    /// as more distinct near-optimal routes (lower `test_route_choice`
-    /// linear%) but also lengthens back-to-back forced-level streaks. Set to
-    /// 1.0 (user decision 2026-07-28: accept the streak hit for the choice).
-    /// Measured: route-choice linear% 34.8→32.5 (2000-seed) at 1.0 vs 0.75;
-    /// forced-level streak 1.79→~1.89 (test_required_progression). Curve is
-    /// non-monotonic for choice — 1.0 is the optimum; 1.25–2.0 regress.
-    /// History: was 0.75 (lowered from 1.5 to cut streaks; streak sweep
-    /// 1.5→2.12, 1.0→1.89, 0.75→1.79, 0.5→1.70, 0.0→1.34). 3.0 clumped hard.
-    pub path_bonus: f64,
-    /// Max detour (BFS hops off the shortest start→target route) that still
-    /// earns any path bonus.
-    pub path_detour_cap: f64,
-    /// A/B experiment: when true, the greedy (aesthetic-max) first half of
-    /// level placement is replaced by uniform random picks — buys back the
-    /// reroll era's raw layout variance at the cost of spread quality. The
-    /// measured second half is unaffected.
-    pub random_first_half: bool,
-}
-
-impl Default for LevelScoring {
-    fn default() -> Self {
-        LevelScoring {
-            spread: SpreadScoring::default(),
-            dead_end_bonus: 0.5,
-            path_bonus: 1.0,
-            path_detour_cap: 6.0,
-            random_first_half: false,
-        }
-    }
-}
-
-/// Knobs for lock placement (`place_locks`). Scores are integer points on a
-/// base of "gated node count" (how much of the map the closed lock walls
-/// off).
-#[derive(Clone, Copy, Debug)]
-pub(crate) struct LockScoring {
-    /// Bonus when the closed lock makes a LATER fort unreachable. Held at 0:
-    /// this bonus is the chain magnet (greedy converges to all-chain worlds
-    /// with it high — measured in the archetype era), and the choice-guard
-    /// re-measure in `place_locks` is its replacement. Kept as a knob for
-    /// sweeps.
-    pub blocks_later_fort_bonus: i32,
-    /// Bonus for the FIRST lock in a world that gates the airship/Bowser.
-    /// Later target-blockers get nothing (they'd pile up next to the
-    /// airship).
-    pub blocks_target_bonus: i32,
-    /// Spread penalty radius: locks closer than this (Manhattan) to an
-    /// already-placed lock are penalized...
-    pub spread_radius: i32,
-    /// ...by this many points per tile of shortfall.
-    pub spread_penalty_per_tile: i32,
-    /// Nudge toward bridge tiles (water gaps read better than path locks).
-    pub bridge_bonus: i32,
-    /// W8 only: extra bias toward the screen-3 showcase bridges. Tuned via a
-    /// 200k-seed sweep: +8 puts ≥1 bridge out in ~99.6% of seeds, two in
-    /// ~30%, all four in ~0.08% (a rare treat).
-    pub w8_bridge_bonus: i32,
-    /// If the best candidate scores below this, prefer a safe lock instead —
-    /// no point spending an impactful-lock slot on a weak chokepoint.
-    pub weak_lock_threshold: i32,
-    /// Max candidate locks per section re-measured by the choice-guard
-    /// before giving up and accepting the top-scored one. The lock pass is
-    /// where route choice is actually manufactured, and this is its budget:
-    /// higher = the guard finds route-creating locks more often (lower
-    /// linear%), at the cost of more route measures per build. Set to 8
-    /// (from 4) — a modest, streak-free, ~+5ms/build choice gain that keeps
-    /// worst-case build latency inside the WASM budget. Sweep (2000-seed
-    /// route-choice linear%, all else default): 4→34.8, 8→33.5, 12→31.8,
-    /// 16→30.0, 24→29.6, but build-time mean climbs 60→74→90ms+ past 12.
-    pub choice_guard_tries: usize,
-    /// Whether some lock MUST sever the goal (the goal-gate duty). Default
-    /// FALSE (user decision 2026-07-27): the shipped guarantee is the C1
-    /// floor, not the door-and-key moment — every lock places purely by
-    /// (choice delta, score) and the goal may be open at start (~1% of
-    /// worlds). The build loop still flips this on per-world as a FLOOR
-    /// BACKSTOP: a world whose ungated cheapest route can't be repaired to
-    /// `C1_FLOOR` (connectivity-pipe corridors that dodge level moves for
-    /// free) is rebuilt gated — the goal-severing lock is the one tool that
-    /// prices such a corridor. See `test_c1_floor_probe`.
-    pub goal_gate: bool,
-}
-
-impl Default for LockScoring {
-    fn default() -> Self {
-        LockScoring {
-            blocks_later_fort_bonus: 0,
-            blocks_target_bonus: 10,
-            spread_radius: 8,
-            spread_penalty_per_tile: 2,
-            bridge_bonus: 1,
-            w8_bridge_bonus: 8,
-            weak_lock_threshold: 5,
-            choice_guard_tries: 8,
-            goal_gate: false,
-        }
-    }
-}
-
-/// Knobs for pipe placement — both the connectivity pass (`place_pipes`) and
-/// the post-lock spare-pipe pass (`place_spare_pipes`).
-#[derive(Clone, Copy, Debug)]
-pub(crate) struct PipeScoring {
-    /// Softmax temperature for every pipe candidate pick. Score range is
-    /// typically ~[-8, +12].
-    pub softmax_t: f64,
-    /// Connectivity build-outward (issue #121): weight on an endpoint's
-    /// walk-degree, so a bridge lands on a junction (a branch node the later
-    /// passes can fork) rather than a dead-end tip (a one-way spur). This is
-    /// the topology-aware endpoint bias — it shapes HOW an island connects,
-    /// never whether (reachability is enforced by the reachable/unreachable
-    /// split, not scored).
-    pub junction_weight: f64,
-    /// Connectivity build-outward: penalty per Manhattan hop of bridge length.
-    /// Keeps the junction bias from reaching back across the map — the island
-    /// still joins via a short bridge to a NEARBY junction, not a distant one.
-    pub bridge_penalty: f64,
-    /// Per-spare-pipe skip-cap weights: relative chance the pipe's max
-    /// level-skip is 1, 2, 3, or 4. Each pipe rolls a cap, then greedily
-    /// takes the biggest skip within it — big skips happen sometimes, not
-    /// every pipe. Tuned via sweep (≈ halves W2/W6 skip size vs uncapped
-    /// while keeping most of the anti-linearity win).
-    pub spare_cap_weights: [f64; 4],
-    /// Spare-pipe score: points per level the pipe lets the player skip.
-    pub level_skip_weight: f64,
-    /// Spare-pipe score: points per hop of route-distance jump (tie-breaker
-    /// under `level_skip_weight`).
-    pub jump_weight: f64,
-    /// Whether a spare pipe may deliberately bridge across one closed lock.
-    /// See [`FortSkipPolicy`]. Rejection of 2+-fort bypasses and goal-opening
-    /// pipes applies under every policy.
-    pub fort_skip: FortSkipPolicy,
-}
-
-impl Default for PipeScoring {
-    fn default() -> Self {
-        PipeScoring {
-            softmax_t: 4.0,
-            junction_weight: 5.0,
-            bridge_penalty: 5.0,
-            spare_cap_weights: [30.0, 25.0, 25.0, 20.0],
-            level_skip_weight: 10.0,
-            jump_weight: 1.0,
-            fort_skip: FortSkipPolicy::AnyOneFort,
-        }
-    }
-}
-
-/// Every placement pass's knobs, bundled. One instance (currently the
-/// defaults) is built per randomization and threaded through `build_world`.
-#[derive(Clone, Copy, Debug, Default)]
-pub(crate) struct Knobs {
-    pub fort: FortScoring,
-    pub level: LevelScoring,
-    pub lock: LockScoring,
-    pub pipe: PipeScoring,
-}
