+++ src/randomize/overworld_build/locks.rs
@@ -1,545 +1,299 @@
-//! Fortress lock / water-gap placement.
+//! Locks phase — one lock (map gap) per fortress, uniform-random placement.
 //!
-//! Per section (in BFS-rank order): score every lockable tile by how much of
-//! the map it walls off, enforce two hard rules (the section's own fort stays
-//! reachable; no earlier fort is stranded), then pick with two choice-first
-//! refinements:
+//! KNOB-FREE: for each fort id, candidate tiles are shuffled and the first
+//! one that keeps the world completable wins. No chokepoint scoring, no
+//! golden-site hunt, no gate duty — the census measures what random locks
+//! gate (including how often the answer is "nothing"), and the shaping
+//! controls get argued from those numbers.
 //!
-//! - **Gate guarantee** — the section named by `gate_section` (the shaping
-//!   pass's random fort pick) hard-filters to candidates that make the goal
-//!   unreachable, so some fort actually gates the goal whenever the geometry
-//!   offers a severing tile. Infeasible sections carry the duty forward.
-//! - **Choice guard** — a lock that would shrink the world's in-band route
-//!   count (measured by `analyze_route_choice`) is demoted and the next-best
-//!   candidate tried, bounded by `choice_guard_tries`. This replaces the
-//!   archetype-era `blocks_later_fort_bonus` chain magnet (now defaulted 0).
+//! Hard invariants (true safety, per the charter):
+//!
+//! - **ORDER-FREE completability** (`WorldState::completable`): close every
+//!   lock, beat every reachable fort, open those locks, repeat — the world
+//!   is valid iff the goal is reached AND every fort is eventually beatable
+//!   (a fort sealed behind its own lock is permanently unplayable content).
+//!   This single fixpoint subsumes the shipping builder's per-section hard
+//!   rules without its "sections beaten in order" linearization.
+//! - **Every fort has a lock.** A fort's Boom-Boom ordinal indexes the world
+//!   FX list, and a lockless fort is dead weight (its beat does nothing). If
+//!   every candidate breaks completability, the fort is REMOVED from the
+//!   world — loudly reported, counted by the census (expected rate ~0: a
+//!   dead-end path tile gates nothing and always passes the fixpoint).
+//! - **Secret-exit safety is computed, not stamped false.** A lock is safe
+//!   iff the world stays completable with that lock sealed forever
+//!   (`completable_sealed`) — the 1-F fortress's secret exit skips the
+//!   lock-opening FX. The ROM-level requirement is at least one safe lock
+//!   ACROSS ALL WORLDS (the writer parks 1-F on a safe slot);
+//!   [`ensure_secret_exit_safe`] is that cross-world backstop.
 
 use super::*;
 
-use super::knobs::LockScoring;
-use super::route_choice::{
-    DEFAULT_SLACK, RouteChoice, SHAPING_SLACK, analyze_route_choice, measure_counts,
-    rescue_targets, walk_edge_mids,
-};
-use super::types::{BuildResult, BuiltWorld, LockAssignment, SlotAssignment, SlotKind, stamp_slots};
+use rand::seq::SliceRandom;
 
-/// The five always-on W8 screen-3 bridges (see `qol::overworld_map`'s
-/// `W8_BRIDGE_EDITS`). Locking one gaps it out as water until its fortress is
-/// beaten — a deliberate showcase, so the lock scorer nudges toward them.
-const W8_BRIDGE_LOCK_POSITIONS: &[(usize, usize)] = &[(5, 51), (5, 53), (5, 55), (5, 57), (5, 59)];
+pub(crate) struct Locks;
 
-/// A scored lock candidate tracked during selection.
-#[derive(Clone, Copy)]
-struct ScoredLock {
-    pos: Pos,
-    gap_tile: u8,
-    replace_tile: u8,
-    score: i32,
-    safe: bool,
-    blocks_target: bool,
+impl Phase for Locks {
+    fn name(&self) -> &'static str {
+        "locks"
+    }
+
+    fn run(&self, state: &mut WorldState, rng: &mut dyn RngCore) -> PhaseReport {
+        let mut actions = Vec::new();
+
+        let fort_ids: Vec<usize> = state
+            .slots
+            .iter()
+            .filter(|s| s.kind == SlotKind::Fortress)
+            .map(|s| s.section)
+            .collect();
+
+        for fort_id in fort_ids {
+            let mut candidates = lock_candidates(state);
+            candidates.shuffle(rng);
+
+            let mut placed = false;
+            let tried = candidates.len();
+            for (pos, tile) in candidates {
+                state.locks.push(LockAssignment {
+                    pos,
+                    gap_tile: gap_tile_for(tile),
+                    replace_tile: tile,
+                    fort_section: fort_id,
+                    // Stamped by recompute_safety_flags once the set is
+                    // final.
+                    secret_exit_safe: false,
+                });
+                if state.completable() {
+                    actions.push(format!("lock for fort {fort_id} at {pos:?}"));
+                    placed = true;
+                    break;
+                }
+                state.locks.pop();
+            }
+            if !placed {
+                // Invariant: every fort has a lock. An unlockable fort is
+                // removed rather than left as dead weight on the map.
+                let idx = state
+                    .slots
+                    .iter()
+                    .position(|s| s.kind == SlotKind::Fortress && s.section == fort_id)
+                    .expect("fort id came from the slot list");
+                let pos = state.slots[idx].pos;
+                state.slots.remove(idx);
+                actions.push(format!(
+                    "fort {fort_id} REMOVED from {pos:?}: all {tried} lock candidates break completability",
+                ));
+            }
+        }
+
+        recompute_safety_flags(state);
+
+        actions.push(format!(
+            "done: {}/{} forts locked, {} secret-exit-safe",
+            state.locks.len(),
+            state.fort_count(),
+            state.locks.iter().filter(|l| l.secret_exit_safe).count(),
+        ));
+        PhaseReport { phase: self.name(), actions }
+    }
 }
 
-// Reason: every argument represents a distinct, independent input to lock
-// placement (geometry, slot list, count, knobs, gate hint, safety flag, RNG).
-// No subset clusters into a meaningful concept — bundling would be a clippy
-// bandage, not a real abstraction.
-#[allow(clippy::too_many_arguments)]
-pub(super) fn place_locks<R: Rng>(
-    grid: &Grid,
-    pipe_pairs: &[TeleportEdge],
-    start_pos: Option<(usize, usize)>,
-    target_pos: Option<(usize, usize)>,
-    slots: &[SlotAssignment],
-    fort_count: usize,
-    knobs: &LockScoring,
-    gate_section: Option<usize>,
-    force_safe: bool,
-    world_idx: usize,
-    rng: &mut R,
-) -> Vec<LockAssignment> {
-    let mut locks: Vec<LockAssignment> = Vec::new();
-    let mut locked_tiles: HashSet<(usize, usize)> = HashSet::new();
+/// Stamp every lock's `secret_exit_safe` flag from the final lock set: safe
+/// iff the world stays completable with that lock sealed forever. Computed
+/// after all placements — sealing interacts with the OTHER locks, so the
+/// flag is only meaningful on the finished set.
+pub(crate) fn recompute_safety_flags(state: &mut WorldState) {
+    for li in 0..state.locks.len() {
+        state.locks[li].secret_exit_safe = state.completable_sealed(Some(li));
+    }
+}
 
-    // Trial world with a given lock list. Locks are modeled as conditional
-    // edges by the scorer, so the raw build grid is correct here.
-    let trial_world = |trial_locks: Vec<LockAssignment>| -> BuiltWorld {
-        BuiltWorld {
-            world_idx,
-            grid: grid.clone(),
-            slots: slots.to_vec(),
-            locks: trial_locks,
-            section_count: fort_count,
-            pipe_pairs: pipe_pairs.to_vec(),
-            hb_sprites: Vec::new(),
-        }
-    };
-    // Full planning measure (wide band, paths populated) — the golden-site
-    // derivation reads route paths.
-    let measure = |trial_locks: Vec<LockAssignment>| -> RouteChoice {
-        analyze_route_choice(&trial_world(trial_locks), SHAPING_SLACK)
-    };
-    // In-band route count — the choice-delta yardstick. Counts-only trial
-    // measure: narrow band, no path reconstruction.
-    let measure_routes = |trial_locks: Vec<LockAssignment>| -> usize {
-        measure_counts(&trial_world(trial_locks), DEFAULT_SLACK).routes.len()
-    };
-    let assignment_for = |c: &ScoredLock, section_idx: usize| LockAssignment {
-        pos: c.pos,
-        gap_tile: c.gap_tile,
-        replace_tile: c.replace_tile,
-        fort_section: section_idx,
-        secret_exit_safe: c.safe,
-        blocks_target: c.blocks_target,
-    };
+/// Cross-world invariant: at least one lock across all worlds must be
+/// secret-exit-safe — the write phase parks the 1-F fortress level (whose
+/// secret exit skips the lock-opening FX) on a slot whose lock can stay
+/// closed forever without softlocking.
+///
+/// Knob-free backstop, run after every world's schedule: if uniform
+/// placement produced no safe lock anywhere, relocate ONE existing lock —
+/// shuffled worlds, shuffled forts, shuffled candidates, first tile that
+/// keeps the world completable both normally and with the new lock sealed.
+pub(crate) fn ensure_secret_exit_safe(
+    worlds: &mut [WorldState],
+    rng: &mut dyn RngCore,
+) -> PhaseReport {
+    let mut actions = Vec::new();
+    let phase = "secret_exit_safety";
 
-    // Build a base grid with forts/levels stamped so BFS sees them as nodes.
-    // This grid does NOT have any locks on it.
-    let mut base_grid = grid.clone();
-    stamp_slots(&mut base_grid, slots);
+    if worlds
+        .iter()
+        .any(|w| w.locks.iter().any(|l| l.secret_exit_safe))
+    {
+        actions.push("already satisfied: uniform placement produced a safe lock".into());
+        return PhaseReport { phase, actions };
+    }
 
-    // Gate duty off (C1-floor experiment arm): drop the hint entirely — every
-    // section places by the normal (choice delta, score) selection below.
-    let gate_section = if knobs.goal_gate { gate_section } else { None };
-
-    // Re-anchor the gate hint on feasibility: a section can host the goal
-    // gate only if SOME severing tile keeps its own fort and every earlier
-    // fort reachable (the hard rules, approximated on the lockless grid).
-    // A blind random hint strands the duty on sections that can never gate
-    // (deep forts sit goal-side of the severs), and the carry-forward then
-    // exhausts — measured as goal-open worlds. Uniform choice among
-    // FEASIBLE sections keeps the no-tell property and the guarantee.
-    let mut gate_pending = if let Some(hint) = gate_section {
-        let fort_of: Vec<Option<Pos>> = (0..fort_count)
-            .map(|sec| {
-                slots
-                    .iter()
-                    .find(|s| s.section == sec && s.kind == SlotKind::Fortress)
-                    .map(|s| s.pos)
-            })
-            .collect();
-        let mut feasible: Vec<usize> = Vec::new();
-        'tiles: {
-            let mut sever_walks: Vec<Reach> = Vec::new();
-            let Some(tp) = target_pos else { break 'tiles };
-            for r in 0..base_grid.rows() {
-                for c in 0..base_grid.cols {
-                    let tile = base_grid.get(r, c);
-                    if !LOCKABLE_TILES.contains(&tile) {
-                        continue;
-                    }
-                    let mut g = base_grid.clone();
-                    g.set(r, c, gap_tile_for(tile));
-                    let nodes = walk_reachable(&g, pipe_pairs, start_pos, world_idx);
-                    if !nodes.contains(tp) {
-                        sever_walks.push(nodes);
-                    }
+    let mut world_order: Vec<usize> = (0..worlds.len()).collect();
+    world_order.shuffle(rng);
+    for wi in world_order {
+        let state = &mut worlds[wi];
+        let mut lock_indices: Vec<usize> = (0..state.locks.len()).collect();
+        lock_indices.shuffle(rng);
+        for li in lock_indices {
+            let original = state.locks[li].clone();
+            let mut candidates = lock_candidates(state);
+            candidates.shuffle(rng);
+            for (pos, tile) in candidates {
+                state.locks[li] = LockAssignment {
+                    pos,
+                    gap_tile: gap_tile_for(tile),
+                    replace_tile: tile,
+                    fort_section: original.fort_section,
+                    secret_exit_safe: false,
+                };
+                if state.completable() && state.completable_sealed(Some(li)) {
+                    // Relocation changed the world — every flag in it is
+                    // stale, not just the moved lock's.
+                    recompute_safety_flags(state);
+                    actions.push(format!(
+                        "W{} fort {} lock relocated {:?} -> {pos:?} (secret-exit-safe)",
+                        state.world_idx + 1,
+                        original.fort_section,
+                        original.pos,
+                    ));
+                    return PhaseReport { phase, actions };
                 }
             }
-            for sec in 0..fort_count {
-                let ok = sever_walks.iter().any(|nodes| {
-                    (0..=sec).all(|j| fort_of[j].is_none_or(|fp| nodes.contains(fp)))
-                });
-                if ok {
-                    feasible.push(sec);
-                }
-            }
+            state.locks[li] = original;
         }
-        // Keep the hint when it's feasible (or nothing provably gates and
-        // the backstops must do what they can); otherwise re-pick uniformly
-        // among the feasible sections.
-        if feasible.contains(&hint) || feasible.is_empty() {
-            Some(hint)
-        } else {
-            Some(feasible[rng.random_range(..feasible.len())])
-        }
-    } else {
-        None
-    };
+    }
 
-    // Process the gate section FIRST, then the rest in section order. The
-    // gate's hard rules are evaluated with no other lock committed (maximum
-    // feasibility — committed safe locks otherwise combine with the sever to
-    // strand forts and kill every gate candidate); the other sections then
-    // place around the committed gate. `build_test_grid` keys progression on
-    // section NUMBERS, not commit order, so the simulation stays correct.
-    let order: Vec<usize> = match gate_pending {
-        Some(g) if g < fort_count => {
-            std::iter::once(g).chain((0..fort_count).filter(|&s| s != g)).collect()
-        }
-        _ => (0..fort_count).collect(),
-    };
-    for (oi, &section_idx) in order.iter().enumerate() {
-        let fort_pos = match slots
-            .iter()
-            .find(|s| s.section == section_idx && s.kind == SlotKind::Fortress)
-        {
-            Some(s) => s.pos,
-            None => {
-                // A fortless section can't lock anything — pass any pending
-                // gate duty along instead of silently dropping it.
-                if gate_pending == Some(section_idx) {
-                    gate_pending = order.get(oi + 1).copied();
-                }
+    actions.push("UNSATISFIED: no relocation in any world yields a safe lock".into());
+    PhaseReport { phase, actions }
+}
+
+/// Which candidates a lock-placement pass admits, strictest first.
+#[derive(Clone, Copy, PartialEq)]
+enum LockPass {
+    /// Closing the tile makes the GOAL unreachable in the all-open world —
+    /// the pipe-proof cut (nothing routes around the goal approach), the
+    /// same duty the shipping builder's C1-floor backstop assigns.
+    GoalGate,
+    /// Closing the tile walls off at least one node (the
+    /// [`WorldState::zero_gate_locks`] definition).
+    AnyGate,
+    /// Any completable tile.
+    Any,
+}
+
+/// Re-place every lock from scratch, preferring candidates that actually
+/// gate something. The shaping loop's wide lock move: clears the lock set
+/// and rebuilds it one fort at a time (shuffled order) — for each fort, the
+/// passes in [`LockPass`] order over shuffled candidates, and every
+/// placement keeps the completability fixpoint true.
+///
+/// `goal_first` (cost mode's request): the first fort that CAN gate the
+/// goal does — a goal gate is the one cut a pipe web cannot bypass, so it
+/// reliably raises C1 on the express-shaped worlds where "gate anything"
+/// finds only worthless cuts. Only one goal gate is sought; the remaining
+/// forts fall back to the normal passes (a goal crowded by every lock is
+/// not a better world).
+///
+/// Returns false if some fort ends up unlockable — the caller owns the
+/// snapshot and must restore it (unlike the dumb Locks phase, shaping never
+/// removes forts; a failed proposal is just rolled back).
+pub(crate) fn place_locks_gating(
+    state: &mut WorldState,
+    rng: &mut dyn RngCore,
+    goal_first: bool,
+) -> bool {
+    state.locks.clear();
+
+    let mut open = state.grid.clone();
+    stamp_slots(&mut open, &state.slots);
+    let open_len = walk_reachable(&open, &state.pipe_pairs, state.start, state.world_idx).len();
+
+    let mut fort_ids: Vec<usize> = state
+        .slots
+        .iter()
+        .filter(|s| s.kind == SlotKind::Fortress)
+        .map(|s| s.section)
+        .collect();
+    fort_ids.shuffle(rng);
+
+    let mut goal_gated = false;
+    for fort_id in fort_ids {
+        let mut candidates = lock_candidates(state);
+        candidates.shuffle(rng);
+
+        let mut placed = false;
+        'passes: for pass in [LockPass::GoalGate, LockPass::AnyGate, LockPass::Any] {
+            if pass == LockPass::GoalGate && (!goal_first || goal_gated) {
                 continue;
             }
-        };
-
-        // Build the "current state" grid: base grid + all previously placed locks
-        // + all locks from earlier sections opened (simulating progression).
-        // When checking section N's lock, sections 0..N-1 have been beaten,
-        // so their locks are open. The new lock we're testing is the only closed one.
-        let build_test_grid = |new_lock: Option<((usize, usize), u8)>| -> Grid {
-            let mut g = base_grid.clone();
-            // Place all previously committed locks
-            for prev in &locks {
-                if prev.fort_section < section_idx {
-                    // Earlier section — fort beaten, lock opened (restore path tile)
-                    g.set(prev.pos.0, prev.pos.1, prev.replace_tile);
-                } else {
-                    // Same or later section — lock still closed
-                    g.set(prev.pos.0, prev.pos.1, prev.gap_tile);
-                }
-            }
-            // Place the candidate lock
-            if let Some((pos, gap)) = new_lock {
-                g.set(pos.0, pos.1, gap);
-            }
-            g
-        };
-
-        // Find all lockable path tiles not yet used
-        let reference_grid = build_test_grid(None);
-        let mut candidates: Vec<(usize, usize)> = Vec::new();
-        for r in 0..reference_grid.rows() {
-            for c in 0..reference_grid.cols {
-                let tile = reference_grid.get(r, c);
-                if LOCKABLE_TILES.contains(&tile) && !locked_tiles.contains(&(r, c)) {
-                    // Row 7 and row 8 share Map_Completions bit ($01).
-                    // A lock/bridge/gap is completion-unsafe — it would
-                    // prevent the fallthrough between rows 7 and 8.
-                    // Skip if the paired row has a completable slot.
-                    if r == 7 || r == 8 {
-                        let paired_row = if r == 7 { 8 } else { 7 };
-                        let pair_completable = slots.iter().any(|s| {
-                            s.pos == (paired_row, c)
-                                && matches!(s.kind, SlotKind::Level | SlotKind::Fortress | SlotKind::Pipe | SlotKind::BonusGame | SlotKind::ToadHouse)
-                        });
-                        if pair_completable {
+            for &(pos, tile) in &candidates {
+                match pass {
+                    LockPass::Any => {}
+                    LockPass::GoalGate | LockPass::AnyGate => {
+                        let mut g = open.clone();
+                        g.set(pos.0, pos.1, gap_tile_for(tile));
+                        let reach =
+                            walk_reachable(&g, &state.pipe_pairs, state.start, state.world_idx);
+                        let admits = match pass {
+                            LockPass::GoalGate => {
+                                state.target.is_some_and(|t| !reach.contains(t))
+                            }
+                            _ => reach.len() < open_len,
+                        };
+                        if !admits {
                             continue;
                         }
                     }
-                    candidates.push((r, c));
                 }
-            }
-        }
-
-        candidates.shuffle(rng);
-
-        // All hard-rule-passing candidates, scored. Selection happens after
-        // the loop (gate filter → safe preference → choice guard).
-        let mut scored: Vec<ScoredLock> = Vec::new();
-
-        // Open grid (no candidate lock) is constant for all candidates in this
-        // section — hoist the BFS to avoid redundant walks per candidate.
-        let open_grid = build_test_grid(None);
-        let open_node_count = walk_reachable(&open_grid, pipe_pairs, start_pos, world_idx).len() as i32;
-
-        // If a previous lock in this world already blocks the target, suppress
-        // the target-blocking bonus to avoid stacking multiple locks against
-        // the airship/Bowser.
-        let target_already_locked = locks.iter().any(|l| l.blocks_target);
-
-        for &cand_pos in &candidates {
-            let tile = reference_grid.get(cand_pos.0, cand_pos.1);
-            let gap = gap_tile_for(tile);
-
-            // Hard rule 1: with this lock placed (and earlier locks opened),
-            // the current fortress must still be reachable from start.
-            let test_grid = build_test_grid(Some((cand_pos, gap)));
-            let walk = walk_reachable(&test_grid, pipe_pairs, start_pos, world_idx);
-
-            if !walk.contains(fort_pos) {
-                continue;
-            }
-
-            // Hard rule 2: this lock must not block any earlier fortress.
-            // Check each earlier section's fort is reachable when its own
-            // lock (and all locks before it) are open but this new lock is closed.
-            let blocks_earlier = locks.iter().any(|prev_lock| {
-                let prev_fort = slots.iter()
-                    .find(|s| s.section == prev_lock.fort_section && s.kind == SlotKind::Fortress);
-                if let Some(pf) = prev_fort {
-                    // Build grid: open locks up to prev_lock's section, close the rest + candidate
-                    let mut g = base_grid.clone();
-                    for l in &locks {
-                        if l.fort_section < prev_lock.fort_section {
-                            g.set(l.pos.0, l.pos.1, l.replace_tile);
-                        } else {
-                            g.set(l.pos.0, l.pos.1, l.gap_tile);
-                        }
+                state.locks.push(LockAssignment {
+                    pos,
+                    gap_tile: gap_tile_for(tile),
+                    replace_tile: tile,
+                    fort_section: fort_id,
+                    secret_exit_safe: false,
+                });
+                if state.completable() {
+                    placed = true;
+                    if pass == LockPass::GoalGate {
+                        goal_gated = true;
                     }
-                    // Also place the candidate lock. Gate-first ordering can
-                    // commit a LATER section before this one, so the candidate
-                    // is open (already beaten) at prev_lock's checkpoint when
-                    // its section is earlier — same rule as the loop above.
-                    let cand_tile = if section_idx < prev_lock.fort_section { tile } else { gap };
-                    g.set(cand_pos.0, cand_pos.1, cand_tile);
-                    let w = walk_reachable(&g, pipe_pairs, start_pos, world_idx);
-                    !w.contains(pf.pos)
-                } else {
-                    false
+                    break 'passes;
                 }
-            });
-            if blocks_earlier {
-                continue;
-            }
-
-            // Check if target is reachable with this lock closed (used for
-            // secret exit safety).
-            let target_reachable = target_pos
-                .map(|tp| walk.contains(tp))
-                .unwrap_or(true);
-
-            // A "safe" lock blocks nothing important: all fortresses and
-            // the target remain reachable. Safe for 1-F secret exit since
-            // leaving it closed can never cause a softlock.
-            let safe = target_reachable && slots.iter().all(|s| {
-                s.kind != SlotKind::Fortress || walk.contains(s.pos)
-            });
-
-            // Score by gated node count: how many nodes become unreachable
-            // when this lock is closed? Prefers chokepoints that gate large
-            // portions of the map over locks adjacent to the airship (which
-            // only gate ~1 node).
-            let gated = open_node_count - walk.len() as i32;
-
-            let mut score: i32 = gated;
-
-            // Bonus: blocks a later fortress (strong progression signal)
-            let blocks_later_fort = slots.iter().any(|s| {
-                s.kind == SlotKind::Fortress
-                    && s.section > section_idx
-                    && !walk.contains(s.pos)
-            });
-            if blocks_later_fort {
-                score += knobs.blocks_later_fort_bonus;
-            }
-
-            // Bonus: blocks the target (airship/bowser) — only credited to
-            // the first such lock in the world; subsequent target-blockers
-            // would just pile up next to the airship.
-            if !target_reachable && !target_already_locked {
-                score += knobs.blocks_target_bonus;
-            }
-
-            // Spread penalty: discourage placing this lock close to any
-            // already-placed lock in the world. Falls off linearly with
-            // Manhattan distance, zero past `spread_radius` tiles.
-            if let Some(min_dist) = locks
-                .iter()
-                .map(|l| cand_pos.0.abs_diff(l.pos.0) + cand_pos.1.abs_diff(l.pos.1))
-                .min()
-            {
-                score -= (knobs.spread_radius - min_dist as i32).max(0)
-                    * knobs.spread_penalty_per_tile;
-            }
-
-            // Slight preference for bridge tiles — water gaps look better
-            // than locks on regular path tiles.
-            if tile == 0xB3 {
-                score += knobs.bridge_bonus;
-            }
-
-            // W8-specific: bias harder toward the screen-3 showcase bridges so
-            // they're gated out more often than raw chokepoint value alone
-            // would pick them.
-            if world_idx == 7 && W8_BRIDGE_LOCK_POSITIONS.contains(&cand_pos) {
-                score += knobs.w8_bridge_bonus;
-            }
-
-            // (A safe lock never blocks the target, so its blocks_target is
-            // false.)
-            scored.push(ScoredLock {
-                pos: cand_pos,
-                gap_tile: gap,
-                replace_tile: tile,
-                score,
-                safe,
-                blocks_target: !target_reachable,
-            });
-        }
-
-        // Sort desc by score. Stable — keeps the shuffled candidate order
-        // among ties, matching the old first-best-wins behavior.
-        scored.sort_by_key(|c| std::cmp::Reverse(c.score));
-        let mut ordered = scored;
-
-        // Prefer safe when forced (retry path) or when the best candidate is
-        // weak anyway — don't spend an impactful-lock slot on a weak
-        // chokepoint. Safe candidates float to the front, order kept within
-        // each half.
-        let best_score = ordered.first().map(|c| c.score).unwrap_or(0);
-        if force_safe || best_score < knobs.weak_lock_threshold {
-            let (safe_first, rest): (Vec<ScoredLock>, Vec<ScoredLock>) =
-                ordered.into_iter().partition(|c| c.safe);
-            ordered = safe_first;
-            ordered.extend(rest);
-        }
-
-        let baseline = measure_routes(locks.clone());
-        // Measure a candidate's effect on the in-band route count and fold
-        // it into a running best by (delta, then score). No negative skip:
-        // max-by-delta already prefers non-degrading candidates, and when
-        // EVERYTHING degrades this picks the least destructive one.
-        let mut best: Option<(i64, i32, ScoredLock)> = None;
-        let measure_into_best = |cand: &ScoredLock, best: &mut Option<(i64, i32, ScoredLock)>| {
-            let mut trial = locks.clone();
-            trial.push(assignment_for(cand, section_idx));
-            let delta = measure_routes(trial) as i64 - baseline as i64;
-            if best.as_ref().is_none_or(|&(d, s, _)| (delta, cand.score) > (d, s)) {
-                *best = Some((delta, cand.score, *cand));
-            }
-        };
-
-        // Gate guarantee: if this section owes the goal a gate, measure the
-        // severing candidates and take the best (delta, score) — a post-merge
-        // trunk tile gates at delta 0 when one exists; otherwise the least
-        // destructive sever wins, because the gate outranks choice. The duty
-        // is carried to the next section only when this one can't sever the
-        // goal at all (the feasibility pre-pass makes that rare).
-        let mut chosen: Option<ScoredLock> = None;
-        if gate_pending == Some(section_idx) {
-            let gating: Vec<ScoredLock> =
-                ordered.iter().copied().filter(|c| c.blocks_target).collect();
-            if gating.is_empty() {
-                // Hand the duty to the next section processed. If none is
-                // left, no section could sever the goal — fall through; the
-                // blocks_target_bonus soft backstop has done what it can.
-                gate_pending = order.get(oi + 1).copied();
-            } else {
-                for cand in gating.iter().take(knobs.choice_guard_tries) {
-                    measure_into_best(cand, &mut best);
-                }
-                if let Some((_, _, c)) = best {
-                    chosen = Some(c);
-                    gate_pending = None;
-                }
+                state.locks.pop();
             }
         }
-
-        // Normal selection: measure a pool — the golden route-derived sites
-        // plus the top-scored candidates — best by (delta, score). force_safe
-        // skips the golden hunt: the retry needs a SAFE lock, not a choice,
-        // and the safe-first ordering plus delta max already serves that.
-        if chosen.is_none() {
-            if ordered.len() <= 1 {
-                chosen = ordered.first().copied();
-            } else {
-                // Golden sites: mid path-tiles of the cheap route's exclusive
-                // walk edges vs each rescue target. Gapping one forces this
-                // fort's +5 onto the shortcut while the target route walks
-                // around — the choice-creating lock. These gate ~0 nodes, so
-                // they never surface in the score ordering; measured
-                // explicitly.
-                let golden: Vec<Pos> = if force_safe {
-                    Vec::new()
-                } else {
-                    let rc = measure(locks.clone());
-                    let mut sites: Vec<Pos> = Vec::new();
-                    if let Some(cheap) = rc.routes.first() {
-                        let cheap_mids = walk_edge_mids(&cheap.path);
-                        for target in rescue_targets(&rc) {
-                            let target_mids = walk_edge_mids(&target.path);
-                            let mut exclusive: Vec<Pos> =
-                                cheap_mids.difference(&target_mids).copied().collect();
-                            exclusive.sort();
-                            sites.extend(exclusive);
-                        }
-                    }
-                    sites
-                };
-                let mut pool: Vec<ScoredLock> = Vec::new();
-                let mut seen: HashSet<Pos> = HashSet::new();
-                for pos in golden {
-                    if let Some(c) = ordered.iter().find(|c| c.pos == pos)
-                        && seen.insert(c.pos)
-                        && pool.len() < 3
-                    {
-                        pool.push(*c);
-                    }
-                }
-                for cand in ordered.iter().take(knobs.choice_guard_tries) {
-                    if seen.insert(cand.pos) {
-                        pool.push(*cand);
-                    }
-                }
-                for cand in pool {
-                    measure_into_best(&cand, &mut best);
-                }
-                chosen = best.map(|(_, _, c)| c).or_else(|| ordered.first().copied());
-            }
-        }
-
-        if let Some(c) = chosen {
-            locked_tiles.insert(c.pos);
-            locks.push(LockAssignment {
-                pos: c.pos,
-                gap_tile: c.gap_tile,
-                replace_tile: c.replace_tile,
-                fort_section: section_idx,
-                secret_exit_safe: c.safe,
-                blocks_target: c.blocks_target,
-            });
+        if !placed {
+            return false;
         }
     }
-
-    // The gate-first processing order commits out of section order; restore
-    // the section-sorted invariant downstream consumers have always seen.
-    locks.sort_by_key(|l| l.fort_section);
-    locks
+    true
 }
 
-/// Stamp build results onto the ROM tile grids for visual inspection.
-///
-/// Writes generic tiles for each slot type so the overworld maps can be
-/// viewed in an emulator. The game will crash if you enter any level.
-#[allow(dead_code)]
-pub(super) fn debug_stamp_rom(rom: &mut crate::rom::Rom, result: &BuildResult) {
-    for built in &result.worlds {
-        let wi = built.world_idx;
-
-        // First write the cleared grid (with pipes already placed)
-        for r in 0..built.grid.rows() {
-            for c in 0..built.grid.cols {
-                let offset = rom_data::map_tile_offset(wi, r, c);
-                rom.data[offset] = built.grid.get(r, c);
+/// Tiles a lock may claim: lockable path-tile types (the kinds the FX engine
+/// can gap out and restore), not already locked, and not the row-7/8 partner
+/// of completable content — a lock consumes that shared completion bit too.
+fn lock_candidates(state: &WorldState) -> Vec<(Pos, u8)> {
+    let locked: HashSet<Pos> = state.locks.iter().map(|l| l.pos).collect();
+    let content: HashSet<Pos> = state.slots.iter().map(|s| s.pos).collect();
+    let mut out = Vec::new();
+    for r in 0..state.grid.rows() {
+        for c in 0..state.grid.cols {
+            let pos = (r, c);
+            let tile = state.grid.get(r, c);
+            if !LOCKABLE_TILES.contains(&tile) || locked.contains(&pos) {
+                continue;
             }
-        }
-
-        // Stamp slot assignments
-        let mut level_num: u8 = 1;
-        for slot in &built.slots {
-            let tile = match slot.kind {
-                SlotKind::Level => {
-                    // Use numbered map tiles ($03-$0D = levels 1-11, then wrap)
-                    let t = 0x02 + level_num.min(13);
-                    level_num = level_num.wrapping_add(1);
-                    t
-                }
-                SlotKind::Fortress => TILE_FORTRESS,
-                SlotKind::Pipe => TILE_PIPE,
-                SlotKind::BonusGame => TILE_BONUS_GAME,
-                SlotKind::ToadHouse => TILE_TOAD_HOUSE,
-                SlotKind::HammerBro => continue, // keep existing blank path tile
-            };
-            let offset = rom_data::map_tile_offset(wi, slot.pos.0, slot.pos.1);
-            rom.data[offset] = tile;
-        }
-
-        // Stamp locks
-        for lock in &built.locks {
-            let offset = rom_data::map_tile_offset(wi, lock.pos.0, lock.pos.1);
-            rom.data[offset] = lock.gap_tile;
+            if let Some(partner) = row78_partner(pos)
+                && content.contains(&partner)
+            {
+                continue;
+            }
+            out.push((pos, tile));
         }
     }
+    out
 }
