+++ src/randomize/overworld_build/mod.rs
@@ -1,81 +1,177 @@
-//! Phase 3 of the overworld pipeline: assign levels to map slots via BFS-ordered
-//! placement, then shape fortresses/locks/pipes against the measured route
-//! structure (choice-first — the builder's job is to hand the player
-//! decisions). Steps live in submodules (types, knobs, scoring, capacity,
-//! sections, shape, pipes, locks, route_choice, progression); this module
-//! holds the `build` entry point that drives them.
+//! Phase 3 of the overworld pipeline: the overworld builder. Takes the
+//! picked-up maps (blank path tiles) and places levels, fortresses, pipes,
+//! locks, and hammer-bro fillers, producing the `BuildResult` the writer
+//! stamps onto the ROM.
+//!
+//! ## Design charter (2026-07-30 session; the "choice-first rebuild")
+//!
+//! - **Tool roles.** The lock is the shaper (a fort on the forced path makes
+//!   its lock decorative — the player opens it regardless). Forts belong off
+//!   the path. Pipes are a primary routing tool, not defensive filler. Level
+//!   moves are a final tweak, not an early shaper.
+//! - **Reorderable phases.** Every placement pass is a [`Phase`]: take the
+//!   [`WorldState`], change it, report what you did. The pipeline is a plain
+//!   list, so the schedule is data — reordering it (even per seed) is a
+//!   caller decision, not a rewrite.
+//! - **Soft preferences over hard rules.** Only true safety is hard
+//!   (completability, secret-exit safety). Notable structures — an ungated
+//!   shortcut, an open goal — are measured rates to be tuned rare, not banned:
+//!   a rare surprise is a good gameplay moment.
+//! - **Knob-free placement, measured judgment.** The placement phases are
+//!   uniform-random; all judgment concentrates in the [`shaping`] loop and
+//!   is justified by the census tables in the test suite.
+//!
+//! ## Pipeline
+//!
+//! Per world: [`Connectivity`] (bridge islands with pipe pairs) →
+//! [`Levels`] → [`Forts`] → [`Locks`] → [`Shaping`] (the diagnosis-driven
+//! improvement loop), wrapped in [`run_shaped_with_web_retries`] (a world
+//! that finishes below the C1 floor redeals its pipe web). Then across
+//! worlds: the secret-exit-safety backstop, hammer-bro fill, toad-house /
+//! spade promotion, and wandering-sprite redistribution.
+//!
+//! Module layout: [`state`] holds the core definitions (world state, phase
+//! unit); [`sources`] loads worlds (pickup input, vanilla ROM reader,
+//! adapter); [`metrics`] is the measuring stick every phase and census
+//! reads through; [`route_choice`] is the underlying route scorer;
+//! [`capacity`] owns budgets and the cross-world promotion post-passes;
+//! [`progression`] is the older required-progression analyzer (censuses).
 
 use std::collections::{HashMap, HashSet};
 
 use rand::Rng;
+use rand::RngCore;
 use rand::seq::{IndexedRandom, SliceRandom};
 
-use super::map_walker::{Reach, walk_degree, walk_map, walk_reachable};
-use super::node_catalog::{NodeCatalog, NodeKind};
-use super::overworld_helpers::{find_target, gap_tile_for, LOCKABLE_TILES};
-use super::overworld_pickup::PickupResult;
 use crate::rom::Rom;
+
+use super::map_walker::{Reach, walk_map, walk_reachable};
+use super::node_catalog::{NodeCatalog, NodeKind};
+use super::overworld_helpers::{LOCKABLE_TILES, find_target, gap_tile_for};
+use super::overworld_pickup::{PickupResult, blank_tile_for};
 use super::rom_data::{
     self, BACKGROUND_TILES, Grid, Pos, TILE_BONUS_GAME, TILE_FORTRESS, TILE_NODE, TILE_PIPE,
     TILE_TOAD_HOUSE, TeleportEdge,
-    VALID_HORZ, VALID_VERT,
 };
 
-mod types;
-mod knobs;
-mod scoring;
 mod capacity;
-mod sections;
-mod shape;
-mod pipes;
+mod connectivity;
+mod forts;
+mod hammer_bros;
+mod levels;
 mod locks;
-mod gated_shortcut;
+mod metrics;
 mod progression;
 mod route_choice;
+mod shaping;
+mod sources;
+#[cfg(test)]
+mod spare_pipes;
+mod state;
+mod types;
 
-use capacity::{SPADE_BUDGET, assign_hb_sprites, promote_hb_slots};
-// pub(crate): v2's budget allotment reuses these verbatim (see below).
-pub(crate) use capacity::{distribute_levels, prepare_capacities, redistribute_fortresses};
-use knobs::Knobs;
-use locks::place_locks;
-pub(crate) use pipes::VANILLA_PIPE_PAIRS;
-pub(crate) use scoring::{LEVEL_SPREAD_EXPONENT, VANILLA_LEVEL_COUNT};
-use sections::build_world;
-pub(crate) use types::CapacityPrep;
-use types::WorldSlotCounts;
-
-// Public API consumed by the randomizer and the overworld writer.
-pub use {types::SlotAssignment, types::SlotKind};
-pub(crate) use sections::bfs_ordered;
-pub(crate) use capacity::RESERVED_DYNAMIC_SLOTS;
-pub(crate) use types::{BuildFlags, BuildResult, BuiltWorld, OverworldData};
-// Progression analysis is exercised only by the test suite today (reserved for a
-// future WASM single-seed dump), so surface it just for `tests`.
 #[cfg(test)]
-pub(crate) use progression::{
-    analyze_required_progression, classify_pipes, dump_required_progression, hammer_skip,
-    island_count, level_adjacency_pairs, start_goal_express_pipe, PipeClass,
-};
-// Choice-first route scorer (separate weighted model — see route_choice.rs).
-#[cfg(test)]
-pub(crate) use route_choice::{analyze_route_choice, dump_route_choice};
-// Measurement surface for the v2 builder harness (`overworld_v2`): the v2
-// world state converts to a `BuiltWorld` so both builders are measured with
-// the same route scorer. Test-only until v2 gains a production caller.
-#[cfg(test)]
-pub(crate) use route_choice::{DEFAULT_SLACK, RouteChoice, SHAPING_SLACK};
-#[cfg(test)]
-pub(crate) use types::{LockAssignment, stamp_slots};
-#[cfg(test)]
-pub(crate) use capacity::fixed_positions_for_world;
-// v2 reuses the shipping allotment machinery (fort redistribution, capacity
-// prep, level distribution) and the C1 floor VERBATIM — same variance rules,
-// same floor value, one source of truth.
-#[cfg(test)]
-pub(crate) use route_choice::C1_FLOOR;
-
+mod builder_tests;
 #[cfg(test)]
 mod tests;
+
+use capacity::{SPADE_BUDGET, assign_hb_sprites, promote_hb_slots};
+
+// Public API consumed by the randomizer, the writer, and the feature
+// modules that post-process the build (hands, troll pipes).
+pub use {types::SlotAssignment, types::SlotKind};
+pub(crate) use capacity::{
+    RESERVED_DYNAMIC_SLOTS, VANILLA_PIPE_PAIRS, bfs_ordered, distribute_levels,
+    fixed_positions_for_world, prepare_capacities, redistribute_fortresses,
+};
+pub(crate) use capacity::{LEVEL_SPREAD_EXPONENT, VANILLA_LEVEL_COUNT};
+pub(crate) use route_choice::{C1_FLOOR, DEFAULT_SLACK, RouteChoice, analyze_route_choice};
+pub(crate) use types::{
+    BuildFlags, BuildResult, BuiltWorld, CapacityPrep, LockAssignment, OverworldData, stamp_slots,
+};
+
+// The phase set and its harness surface.
+pub(crate) use connectivity::Connectivity;
+pub(crate) use forts::Forts;
+pub(crate) use hammer_bros::HammerBroFill;
+pub(crate) use levels::Levels;
+pub(crate) use locks::{Locks, ensure_secret_exit_safe};
+pub(crate) use metrics::measure_world;
+pub(crate) use shaping::Shaping;
+pub(crate) use sources::{allot_budgets, from_pickup};
+pub(crate) use state::{Phase, PhaseReport, WorldState, row78_partner, run_schedule};
+
+// Test-only measurement surface: the census/probe harness in the test
+// modules and the diagnostic dumps.
+#[cfg(test)]
+pub(crate) use progression::{
+    PipeClass, analyze_required_progression, classify_pipes, dump_required_progression,
+    hammer_skip, island_count, level_adjacency_pairs, start_goal_express_pipe,
+};
+#[cfg(test)]
+pub(crate) use route_choice::{SHAPING_SLACK, dump_route_choice};
+#[cfg(test)]
+pub(crate) use sources::{from_built, from_vanilla};
+#[cfg(test)]
+pub(crate) use spare_pipes::SparePipes;
+
+/// Pipe-web redeals allowed beyond the first attempt when the finished
+/// world ends below the C1 floor. Retries fire only on the few percent of
+/// worlds that finish sub-floor, so the cost is a handful of extra shaped
+/// runs per hundred seeds — census-watched, like everything else.
+pub(crate) const WEB_RETRIES: usize = 4;
+
+/// The full shaped pipeline for one world, with pipe-web redeals — the
+/// answer to the sub-floor tail (2026-07-31 diagnosis). A world that
+/// finishes below the C1 floor is almost always one unlucky connectivity
+/// draw: a bridge mouth cheap from the start turns the goal into a 1-2 pipe
+/// express, and no shaping move can price a trunk that short (the
+/// SAS-swapped W4/W7 goals — a former start pocket, two tiny walk-arms, two
+/// mandatory bridges — are where the draw goes bad ~1 seed in 9). No
+/// placement bias and no repair move: restore the placement, pick up the
+/// pipe web ONLY (levels and forts keep their spots — layout diversity is
+/// the scarcest currency), and redeal connectivity + locks + shaping on a
+/// fresh uniform web. Screen the outcome, not the proposal. If every
+/// attempt ends sub-floor, the best by (C1, routes in band) is kept.
+pub(crate) fn run_shaped_with_web_retries(state: &mut WorldState, rng: &mut dyn RngCore) {
+    run_schedule(state, &[&Connectivity, &Levels, &Forts], rng);
+    let placement = state.snapshot();
+    run_schedule(state, &[&Locks, &Shaping], rng);
+
+    let mut best: Option<(u32, usize, state::WorldSnapshot)> = None;
+    let mut redeals = 0usize;
+    loop {
+        let m = measure_world(state);
+        if m.c1 >= C1_FLOOR || redeals >= WEB_RETRIES {
+            if let Some((c1, routes, snap)) = &best
+                && (m.c1, m.routes_in_band) < (*c1, *routes)
+            {
+                state.restore(snap);
+            }
+            if redeals > 0 {
+                let kept = measure_world(state);
+                state.log.push(PhaseReport {
+                    phase: "web_retry",
+                    actions: vec![format!(
+                        "{redeals} redeal(s): kept C1 {}, routes {}",
+                        kept.c1, kept.routes_in_band,
+                    )],
+                });
+            }
+            return;
+        }
+        if best
+            .as_ref()
+            .is_none_or(|(c1, routes, _)| (m.c1, m.routes_in_band) > (*c1, *routes))
+        {
+            best = Some((m.c1, m.routes_in_band, state.snapshot()));
+        }
+        redeals += 1;
+        state.restore(&placement);
+        state.pickup_pipes();
+        run_schedule(state, &[&Connectivity, &Locks, &Shaping], rng);
+    }
+}
 
 /// Execute Phase 3: build slot assignments for all 8 worlds.
 pub(crate) fn build<R: Rng>(
@@ -84,134 +180,37 @@ pub(crate) fn build<R: Rng>(
     rng: &mut R,
     flags: BuildFlags,
 ) -> BuildResult {
-    let BuildFlags {
-        shuffle_toad_houses,
-        eights_are_wild,
-        shuffle_hammer_bros,
-    } = flags;
-    let pickup = data.pickup;
-    let catalog = data.catalog;
-    // Step 0: redistribute fortresses
-    let fort_counts = redistribute_fortresses(rng);
+    // Per-seed budgets: fortresses via redistribute_fortresses (W8 keeps 4;
+    // W1-W7 roll 1-3 with the 13-fort total conserved), levels distributed
+    // by compressed capacity shares.
+    let (level_counts, fort_counts) = allot_budgets(rom, data.catalog, data.pickup, &flags, rng);
 
-    // Pre-compute patched grids, fixed positions, and per-world level capacity.
-    // (Shared with the distribution-tuning tests so they can't drift.)
-    let CapacityPrep {
-        patched_grids,
-        fixed_positions,
-        capacities,
-    } = prepare_capacities(
-        rom, catalog, pickup, &fort_counts,
-        eights_are_wild, shuffle_toad_houses, shuffle_hammer_bros,
-    );
+    let mut states: Vec<WorldState> = (0..8)
+        .map(|wi| {
+            let mut state = from_pickup(rom, data.catalog, data.pickup, wi, &flags);
+            state.level_budget = level_counts[wi];
+            state.fort_budget = fort_counts[wi];
+            state
+        })
+        .collect();
 
-    // Distribute VANILLA_LEVEL_COUNT levels across worlds by compressed capacity
-    // (see LEVEL_SPREAD_EXPONENT). The compression keeps the densest worlds from
-    // hoarding levels, so the old W6-specific clamp is no longer needed.
-    let level_counts = distribute_levels(&capacities, VANILLA_LEVEL_COUNT, LEVEL_SPREAD_EXPONENT, rng);
-
-    let knobs = Knobs::default();
-    let mut worlds = Vec::with_capacity(8);
-    for wi in 0..8 {
-        // Max non-pipe slots = pointer table slots minus pipe endpoints.
-        // This caps the total fort+level+HB entries to what the pointer table
-        // can actually hold. Excess blank tiles stay as path nodes.
-        let ptr_slots = pickup.worlds[wi].pool_indices.len();
-        let pipe_endpoints = VANILLA_PIPE_PAIRS[wi] * 2;
-        let max_non_pipe_slots = ptr_slots.saturating_sub(pipe_endpoints);
-
-        let counts = WorldSlotCounts {
-            fort_count: fort_counts[wi],
-            level_count: level_counts[wi],
-            pipe_pair_count: VANILLA_PIPE_PAIRS[wi],
-            max_non_pipe_slots,
-            force_safe: false,
-        };
-        let built = build_world(
-            wi,
-            rom,
-            patched_grids[wi].clone(),
-            &fixed_positions[wi],
-            &counts,
-            &knobs,
-            shuffle_hammer_bros,
-            rng,
-        );
-
-        // C1-floor backstop: with the goal-gate duty off, ~3% of worlds end
-        // below the floor because their cheapest route rides connectivity
-        // pipes — a corridor level moves can't charge (the walk just dodges
-        // any level for free). A goal-severing lock is the one tool that
-        // prices such a corridor, so rebuild ONLY those worlds with the gate
-        // on. Cheap: one counts-only measure per world, a rebuild for the
-        // deficient few.
-        let built = {
-            let rc = route_choice::measure_counts(&built, route_choice::DEFAULT_SLACK);
-            if knobs.lock.goal_gate || !rc.reachable || rc.best_cost >= route_choice::C1_FLOOR {
-                built
-            } else {
-                let mut gated = knobs;
-                gated.lock.goal_gate = true;
-                build_world(
-                    wi,
-                    rom,
-                    patched_grids[wi].clone(),
-                    &fixed_positions[wi],
-                    &counts,
-                    &gated,
-                    shuffle_hammer_bros,
-                    rng,
-                )
-            }
-        };
-        worlds.push(built);
+    for state in &mut states {
+        run_shaped_with_web_retries(state, rng);
     }
 
-    // Ensure at least one lock across all worlds is secret-exit-safe.
-    // The score-based prefer_safe usually produces one (triggers when
-    // best_score < 5), but if not, retry with force_safe=true.
-    let has_safe = worlds.iter().any(|b| b.locks.iter().any(|l| l.secret_exit_safe));
-    if !has_safe {
-        let mut retry_order: Vec<usize> = (0..8).collect();
-        retry_order.shuffle(rng);
-        for &wi in &retry_order {
-            let built = &worlds[wi];
-            let start_pos = rom_data::find_start(&built.grid);
-            let target_pos = find_target(&built.grid, wi);
-            // Retry aims to surface a secret-exit-safe lock: force_safe
-            // floats safe candidates first. The world's goal gate must NOT
-            // be flattened in the process — carry its section through so the
-            // rebuild keeps the goal gated while the other sections go safe.
-            let gate_section = built
-                .locks
-                .iter()
-                .find(|l| l.blocks_target)
-                .map(|l| l.fort_section);
-            // place_locks drops the hint unless goal_gate is on, so a world
-            // gated by the C1-floor backstop must keep gate duty here too.
-            let mut retry_knobs = knobs.lock;
-            if gate_section.is_some() {
-                retry_knobs.goal_gate = true;
-            }
-            let new_locks = place_locks(
-                &built.grid,
-                &built.pipe_pairs,
-                start_pos,
-                target_pos,
-                &built.slots,
-                fort_counts[wi],
-                &retry_knobs,
-                gate_section,
-                true, // force_safe
-                wi,
-                rng,
-            );
-            if new_locks.iter().any(|l| l.secret_exit_safe) {
-                worlds[wi].locks = new_locks;
-                break;
-            }
-        }
+    // Cross-world invariant: at least one lock somewhere must be
+    // secret-exit-safe — the writer parks the 1-F fortress level (whose
+    // secret exit skips the lock-opening FX) on it.
+    let _ = ensure_secret_exit_safe(&mut states, rng);
+
+    // Every leftover reachable blank becomes a HammerBro slot: the pointer
+    // table pool the writer fills, and the promotion stock below.
+    for state in &mut states {
+        run_schedule(state, &[&HammerBroFill], rng);
+        renumber_fort_sections(state);
     }
+
+    let mut worlds: Vec<BuiltWorld> = states.iter().map(|s| s.to_built()).collect();
 
     // Toad houses promote first so the smaller, less flexible 22-entry budget
     // lands before spades scramble for the remaining HammerBro slots.
@@ -228,10 +227,31 @@ pub(crate) fn build<R: Rng>(
     // 1-3 per world, summing to the vanilla total). Runs last so the HammerBro
     // slots it selects from are final (Toad House / spade promotion already
     // consumed any it needed). Decided here; the writer stamps the ROM tables.
-    if shuffle_hammer_bros {
+    if flags.shuffle_hammer_bros {
         assign_hb_sprites(rom, data.pickup, &mut worlds, rng);
     }
 
     BuildResult { worlds, fort_counts }
 }
 
+/// The writer pairs each lock to its fortress by INDEX: fort `section`
+/// values must be a dense `0..section_count`. Placement assigns them
+/// densely, but the Locks phase may remove an unlockable fort (rare),
+/// leaving a hole — renumber in placement order and remap the locks.
+fn renumber_fort_sections(state: &mut WorldState) {
+    let remap: HashMap<usize, usize> = state
+        .slots
+        .iter()
+        .filter(|s| s.kind == SlotKind::Fortress)
+        .enumerate()
+        .map(|(new, s)| (s.section, new))
+        .collect();
+    for slot in &mut state.slots {
+        if slot.kind == SlotKind::Fortress {
+            slot.section = remap[&slot.section];
+        }
+    }
+    for lock in &mut state.locks {
+        lock.fort_section = remap[&lock.fort_section];
+    }
+}
