+++ src/randomize/overworld_build/tests.rs
@@ -3,10 +3,10 @@ use super::capacity::{
     W8_HB_CAP, distribute_levels, fixed_positions_for_world, prepare_capacities,
     redistribute_fortresses,
 };
-use super::locks::debug_stamp_rom;
-use super::pipes::VANILLA_PIPE_PAIRS;
-use super::scoring::{VANILLA_LEVEL_COUNT, is_dead_end};
-use super::sections::find_blank_slots;
+
+
+
+use super::capacity::find_blank_slots;
 use super::types::stamp_slots;
 use crate::rom::Rom;
 use rand::SeedableRng;
@@ -210,12 +210,16 @@ fn test_build_all_worlds() {
         let pipes = built.pipe_pairs.len();
         let locks = built.locks.len();
 
-        assert_eq!(forts, result.fort_counts[wi],
-            "W{}: fort slots {} != expected {}", wi + 1, forts, result.fort_counts[wi]);
-        assert_eq!(pipes, VANILLA_PIPE_PAIRS[wi],
-            "W{}: pipe pairs {} != expected {}", wi + 1, pipes, VANILLA_PIPE_PAIRS[wi]);
-        assert!(locks <= result.fort_counts[wi],
-            "W{}: locks {} > fort count {}", wi + 1, locks, result.fort_counts[wi]);
+        // Forts: the allotment is a target, and the Locks phase may remove an
+        // unlockable fort (rare). Every surviving fort keeps exactly one lock.
+        assert!(forts <= result.fort_counts[wi],
+            "W{}: fort slots {} > allotted {}", wi + 1, forts, result.fort_counts[wi]);
+        // Pipes are a routing tool with a budget, not a quota — the builder
+        // spends as few as connectivity needs plus what shaping earns.
+        assert!(pipes <= VANILLA_PIPE_PAIRS[wi],
+            "W{}: pipe pairs {} > budget {}", wi + 1, pipes, VANILLA_PIPE_PAIRS[wi]);
+        assert_eq!(locks, forts,
+            "W{}: every fort must have a lock ({} locks, {} forts)", wi + 1, locks, forts);
     }
 
     let total_levels: usize = result.worlds.iter()
@@ -226,7 +230,9 @@ fn test_build_all_worlds() {
         .sum();
     assert_eq!(total_levels, VANILLA_LEVEL_COUNT,
         "total levels {} != {}", total_levels, VANILLA_LEVEL_COUNT);
-    assert_eq!(total_forts, 17, "total forts {} != 17", total_forts);
+    // 17 allotted; the Locks phase may remove an unlockable fort (rare).
+    assert!(total_forts <= 17, "total forts {} > 17", total_forts);
+    assert!(total_forts >= 15, "total forts {} suspiciously low", total_forts);
 }
 
 /// Regression: the overworld builder must never strand a world's target
@@ -290,38 +296,6 @@ fn all_world_targets_reachable() {
             }
         }
     }
-}
-
-/// Every built world must stay completable via PROGRESSION — you can beat the
-/// forts in reachable order, opening their locks, and reach the goal. The
-/// all-locks-open connectivity check above misses a lock that strands its own
-/// gate; this uses the authoritative progression oracle instead. Guards the
-/// gated-shortcut move (issue #125), whose added pipe + re-hunted lock are the
-/// one place a builder pass could newly strand the goal, across the realistic
-/// census flag mix (both hammer arms).
-#[test]
-fn gated_shortcut_progression_never_strands_goal() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let failures = par_seeds(200, |seed| {
-        let result = census_build(&rom, seed);
-        let mut bad: Vec<(u64, usize, bool)> = Vec::new();
-        for built in &result.worlds {
-            for hammer in [false, true] {
-                if !analyze_required_progression(built, hammer).reachable {
-                    bad.push((seed, built.world_idx, hammer));
-                }
-            }
-        }
-        bad
-    });
-    let all: Vec<_> = failures.into_iter().flatten().collect();
-    assert!(
-        all.is_empty(),
-        "worlds unreachable via progression (seed, world_idx, hammer): {all:?}",
-    );
 }
 
 /// Tuning diagnostic: sweep the level-spread exponent and show the resulting
@@ -559,91 +533,6 @@ fn hammer_bro_redistribution_invariants() {
     }
 }
 
-#[test]
-fn test_locks_dont_block_own_fort() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-
-    for seed in 0..10 {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let result = build(&rom, &OverworldData { pickup: &pickup, catalog: &catalog }, &mut rng, BuildFlags { shuffle_toad_houses: true, ..Default::default() });
-
-        for built in &result.worlds {
-            let start_pos = rom_data::find_start(&built.grid);
-
-            // Build grid with all assignments stamped (same as production)
-            let mut test_grid = built.grid.clone();
-            stamp_slots(&mut test_grid, &built.slots);
-
-            // For each lock, verify its fort is still reachable
-            for lock in &built.locks {
-                // Place ALL locks
-                let mut locked_grid = test_grid.clone();
-                for l in &built.locks {
-                    locked_grid.set(l.pos.0, l.pos.1, l.gap_tile);
-                }
-                // But open THIS lock (as if its fort was beaten)
-                locked_grid.set(lock.pos.0, lock.pos.1, lock.replace_tile);
-
-                // Open all locks from earlier sections too
-                for earlier in &built.locks {
-                    if earlier.fort_section < lock.fort_section {
-                        locked_grid.set(earlier.pos.0, earlier.pos.1, earlier.replace_tile);
-                    }
-                }
-
-                let fort_pos = built.slots.iter()
-                    .find(|s| s.section == lock.fort_section && s.kind == SlotKind::Fortress)
-                    .map(|s| s.pos);
-
-                if let Some(fp) = fort_pos {
-                    let walk = walk_map(&locked_grid, &built.pipe_pairs, start_pos, built.world_idx);
-                    assert!(walk.nodes.contains(&fp),
-                        "Seed {seed} W{}: lock at {:?} blocks its own fort at {:?}",
-                        built.world_idx + 1, lock.pos, fp);
-                }
-            }
-        }
-    }
-}
-
-#[test]
-#[ignore]
-fn test_dump_debug_rom() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-
-    for seed in [42, 123, 999] {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let result = build(&rom, &OverworldData { pickup: &pickup, catalog: &catalog }, &mut rng, BuildFlags { shuffle_toad_houses: true, ..Default::default() });
-
-        let mut rom_copy = Rom::from_bytes(&rom.data).unwrap();
-        debug_stamp_rom(&mut rom_copy, &result);
-
-        let filename = format!("debug_build_seed{seed}.nes");
-        std::fs::write(&filename, &rom_copy.data).unwrap();
-
-        eprintln!("\n=== Seed {seed} ===");
-        for built in &result.worlds {
-            let forts = built.slots.iter().filter(|s| s.kind == SlotKind::Fortress).count();
-            let levels = built.slots.iter().filter(|s| s.kind == SlotKind::Level).count();
-            let pipes = built.pipe_pairs.len();
-            let locks = built.locks.len();
-            eprintln!(
-                "  W{}: {} forts, {} levels, {} pipe pairs, {} locks",
-                built.world_idx + 1, forts, levels, pipes, locks,
-            );
-        }
-        eprintln!("  Wrote {filename}");
-    }
-}
-
 /// Diagnostic: how often does each world end up with 1, 2, or 3 fortresses?
 /// Runs the real build for 1000 seeds and tallies the actual placed Fortress
 /// slots per world. W8 is fixed at 4. Run with:
@@ -701,173 +590,6 @@ fn fort_count_distribution() {
     for (c, &count) in agg.iter().enumerate().take(4).skip(1) {
         eprintln!("  {} fort(s): {:5} ({:5.1}%)", c, count, 100.0 * count as f64 / total);
     }
-}
-
-/// Least-*levels* distance from `source` to every reachable node: entering a
-/// `blocking` node (a level or fortress the player must clear to pass) costs 1,
-/// everything else 0. 0-1 BFS over the walk graph. The source itself is free.
-/// Canoe reachability is recomputed from `source`, so mid-map sources in canoe
-/// worlds are a slight approximation.
-fn least_levels_from(
-    grid: &Grid,
-    pipe_pairs: &[TeleportEdge],
-    source: Pos,
-    wi: usize,
-    blocking: &HashSet<Pos>,
-) -> HashMap<Pos, usize> {
-    use std::collections::VecDeque;
-    let walk = walk_map(grid, pipe_pairs, Some(source), wi);
-    let mut dist: HashMap<Pos, usize> = HashMap::new();
-    let mut dq: VecDeque<Pos> = VecDeque::new();
-    dist.insert(source, 0);
-    dq.push_back(source);
-    while let Some(u) = dq.pop_front() {
-        let du = dist[&u];
-        if let Some(edges) = walk.edges.get(&u) {
-            for e in edges {
-                let w = usize::from(blocking.contains(&e.dest));
-                let nd = du + w;
-                if nd < *dist.get(&e.dest).unwrap_or(&usize::MAX) {
-                    dist.insert(e.dest, nd);
-                    if w == 0 {
-                        dq.push_front(e.dest);
-                    } else {
-                        dq.push_back(e.dest);
-                    }
-                }
-            }
-        }
-    }
-    dist
-}
-
-/// Pipe diagnostics: per-world pipe-pair counts, how many levels the pipe set
-/// shaves off the least-levels path to the goal (the "level skip" value), and
-/// how often a pipe makes an otherwise-mandatory fort optional (the "fort
-/// skip"). Baseline before the lock-aware spare-pipe pass was ~0 fort skips
-/// (2026-07: 1 of 3735 forts); the fort-skip pass should raise it. Run with:
-///   cargo test --lib pipe_metrics -- --ignored --nocapture
-#[test]
-#[ignore]
-fn pipe_metrics() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-    const SEEDS: u64 = 1000;
-
-    let mut pairs_hist = [0u64; 12]; // pipe-pair count per world
-    let mut worlds_with_pipes = 0u64;
-    let mut conn_required = 0u64; // goal unreachable on foot (pipes needed for access)
-    let mut shortcut_worlds = 0u64; // goal reachable on foot; pipes are pure shortcuts
-    let mut saved_hist = [0u64; 12]; // levels the pipe set saves (shortcut worlds)
-    let mut skipped_forts = 0u64; // forts mandatory on foot but optional with pipes
-    let mut forts_examined = 0u64; // forts in foot-reachable pipe worlds
-    let mut worlds_fort_skip = 0u64;
-
-    for seed in 0..SEEDS {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let result = build(
-            &rom,
-            &OverworldData { pickup: &pickup, catalog: &catalog },
-            &mut rng,
-            BuildFlags { shuffle_toad_houses: true, ..Default::default() },
-        );
-
-        for built in &result.worlds {
-            let wi = built.world_idx;
-            let mut base = built.grid.clone();
-            stamp_slots(&mut base, &built.slots);
-            let Some(start) = rom_data::find_start(&base) else { continue };
-            let Some(target) = find_target(&base, wi) else { continue };
-
-            let npairs = built.pipe_pairs.len();
-            pairs_hist[npairs.min(11)] += 1;
-            if npairs == 0 {
-                continue;
-            }
-            worlds_with_pipes += 1;
-
-            let blocking: HashSet<Pos> = built
-                .slots
-                .iter()
-                .filter(|s| matches!(s.kind, SlotKind::Level | SlotKind::Fortress))
-                .map(|s| s.pos)
-                .collect();
-
-            // Least-levels to goal with all locks OPEN (pure geography), with the
-            // pipe set vs with no pipes at all.
-            let with = least_levels_from(&base, &built.pipe_pairs, start, wi, &blocking);
-            let without = least_levels_from(&base, &[], start, wi, &blocking);
-            let foot_reachable = without.contains_key(&target);
-            if !foot_reachable {
-                conn_required += 1;
-                continue; // can't separate a pipe "skip" from required access
-            }
-            shortcut_worlds += 1;
-            let saved = without[&target].saturating_sub(*with.get(&target).unwrap_or(&without[&target]));
-            saved_hist[saved.min(11)] += 1;
-
-            // Fort skip: with each lock closed (others open), is the fort
-            // mandatory on foot but bypassed once pipes are included?
-            let all_sections: HashSet<usize> = built.locks.iter().map(|l| l.fort_section).collect();
-            let grid_with = |opened: &HashSet<usize>| -> Grid {
-                let mut g = base.clone();
-                for l in &built.locks {
-                    if opened.contains(&l.fort_section) {
-                        g.set(l.pos.0, l.pos.1, l.replace_tile);
-                    } else {
-                        g.set(l.pos.0, l.pos.1, l.gap_tile);
-                    }
-                }
-                g
-            };
-            let mut world_skips = 0u64;
-            for sec in &all_sections {
-                forts_examined += 1;
-                let mut others = all_sections.clone();
-                others.remove(sec);
-                let g = grid_with(&others);
-                let mand_pipes = !walk_map(&g, &built.pipe_pairs, Some(start), wi).nodes.contains(&target);
-                let mand_foot = !walk_map(&g, &[], Some(start), wi).nodes.contains(&target);
-                if mand_foot && !mand_pipes {
-                    skipped_forts += 1;
-                    world_skips += 1;
-                }
-            }
-            if world_skips >= 1 {
-                worlds_fort_skip += 1;
-            }
-        }
-    }
-
-    eprintln!("\n=== Pipe pairs per world ({SEEDS} seeds, 8000 worlds) ===");
-    let ptot: u64 = pairs_hist.iter().sum();
-    for (n, &c) in pairs_hist.iter().enumerate() {
-        if c == 0 {
-            continue;
-        }
-        let label = if n == 11 { "11+".to_string() } else { n.to_string() };
-        eprintln!("  {label:>3} pairs: {c:6} ({:5.1}%)", 100.0 * c as f64 / ptot as f64);
-    }
-
-    eprintln!("\n=== Level skip (worlds with >=1 pipe: {worlds_with_pipes}) ===");
-    eprintln!("  goal needs pipes for access (connectivity): {conn_required} ({:5.1}%)", 100.0 * conn_required as f64 / worlds_with_pipes as f64);
-    eprintln!("  pipes are pure shortcuts (foot-reachable goal): {shortcut_worlds} ({:5.1}%)", 100.0 * shortcut_worlds as f64 / worlds_with_pipes as f64);
-    eprintln!("  levels saved by the pipe set (shortcut worlds):");
-    let stot: u64 = saved_hist.iter().sum();
-    for (s, &c) in saved_hist.iter().enumerate() {
-        if c == 0 {
-            continue;
-        }
-        let label = if s == 11 { "11+".to_string() } else { s.to_string() };
-        eprintln!("    saves {label:>3}: {c:6} ({:5.1}%)", 100.0 * c as f64 / stot as f64);
-    }
-
-    eprintln!("\n=== Fort skip ({forts_examined} forts in foot-reachable pipe worlds) ===");
-    eprintln!("  forts mandatory on foot but bypassed by a pipe: {skipped_forts} ({:5.2}%)", 100.0 * skipped_forts as f64 / forts_examined.max(1) as f64);
-    eprintln!("  worlds with >=1 pipe-skipped fort: {worlds_fort_skip} ({:5.2}% of pipe worlds)", 100.0 * worlds_fort_skip as f64 / worlds_with_pipes as f64);
 }
 
 /// Per-world topology snapshot for the goal-open probes. Distilled from the
@@ -938,765 +660,7 @@ fn world_topology(built: &BuiltWorld) -> Option<WorldTopology> {
     Some(WorldTopology { fort_count: forts.len(), depth })
 }
 
-/// TEMP probe: dump full detail for every world where the goal is reachable
-/// with ALL locks closed (the rare "goal open at start" census bucket).
-#[test]
-#[ignore]
-fn goal_open_probe() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-    let results = par_seeds(1000, |seed| {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        build(
-            &rom,
-            &OverworldData { pickup: &pickup, catalog: &catalog },
-            &mut rng,
-            BuildFlags { shuffle_toad_houses: true, ..Default::default() },
-        )
-    });
-    for (seed, result) in results.iter().enumerate() {
-        let hit = result.worlds.iter().any(|b| {
-            world_topology(b).is_some_and(|t| t.fort_count >= 2 && t.depth == 0)
-        });
-        if !hit {
-            continue;
-        }
-        eprintln!("seed {seed}: full-seed picture (goal-open world marked *)");
-        for built in &result.worlds {
-            let t = world_topology(built);
-            let open = t.as_ref().is_some_and(|t| t.fort_count >= 2 && t.depth == 0);
-            let safe_locks = built.locks.iter().filter(|l| l.secret_exit_safe).count();
-            eprintln!(
-                "  {}W{}: forts {}, safe locks {}/{}",
-                if open { "*" } else { " " },
-                built.world_idx + 1,
-                built.locks.len(),
-                safe_locks,
-                built.locks.len(),
-            );
-            // For the goal-open world: is the goal gateable AT ALL — does any
-            // single lockable tile, gapped, sever the target (pipes+canoes
-            // considered)?
-            if open {
-                let wi = built.world_idx;
-                let mut base = built.grid.clone();
-                stamp_slots(&mut base, &built.slots);
-                let start = rom_data::find_start(&base).unwrap();
-                let target = find_target(&base, wi).unwrap();
-                let mut gateable = 0usize;
-                for r in 0..base.rows() {
-                    for c in 0..base.cols {
-                        let tile = base.get(r, c);
-                        if !LOCKABLE_TILES.contains(&tile) {
-                            continue;
-                        }
-                        let mut g = base.clone();
-                        g.set(r, c, gap_tile_for(tile));
-                        if !walk_map(&g, &built.pipe_pairs, Some(start), wi)
-                            .nodes
-                            .contains(&target)
-                        {
-                            gateable += 1;
-                        }
-                    }
-                }
-                eprintln!(
-                    "     -> single-lock-gateable candidate tiles for the goal: {gateable}"
-                );
-            }
-        }
-    }
-}
 
-/// How close do goal-blocking locks sit to the goal, and how much deeper
-/// COULD they sit if forts weren't in the way? For every world with a
-/// realized target-blocking lock: its Manhattan distance to the target and
-/// its gated-node count, vs the deepest fort-free severing tile available
-/// on that map (the placement headroom). Answers "are doorstep gates a
-/// lock-scoring problem or a fort-placement problem?"
-#[test]
-#[ignore]
-fn goal_gate_depth_probe() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-    let mut dist_hist = [0usize; 12]; // Manhattan distance, last bucket = 11+
-    let mut chosen_gated: Vec<usize> = Vec::new();
-    let mut best_gated: Vec<usize> = Vec::new();
-    let mut at_headroom = 0usize; // chosen gate is within 2 nodes of best available
-    let mut worlds_with_gate = 0usize;
-    for seed in 0..300u64 {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let result = build(
-            &rom,
-            &OverworldData { pickup: &pickup, catalog: &catalog },
-            &mut rng,
-            BuildFlags { shuffle_toad_houses: true, ..Default::default() },
-        );
-        for built in &result.worlds {
-            let wi = built.world_idx;
-            let Some(gate) = built.locks.iter().find(|l| l.blocks_target) else {
-                continue;
-            };
-            let mut base = built.grid.clone();
-            stamp_slots(&mut base, &built.slots);
-            let start = rom_data::find_start(&base);
-            let Some(target) = find_target(&base, wi) else { continue };
-            worlds_with_gate += 1;
-
-            let d = gate.pos.0.abs_diff(target.0) + gate.pos.1.abs_diff(target.1);
-            dist_hist[d.min(11)] += 1;
-
-            let open_nodes = walk_map(&base, &built.pipe_pairs, start, wi).nodes;
-            let forts: Vec<(usize, usize)> = built
-                .slots
-                .iter()
-                .filter(|s| s.kind == SlotKind::Fortress)
-                .map(|s| s.pos)
-                .collect();
-
-            // Chosen gate's gated-node count.
-            let mut g = base.clone();
-            g.set(gate.pos.0, gate.pos.1, gate.gap_tile);
-            let chosen =
-                open_nodes.len() - walk_map(&g, &built.pipe_pairs, start, wi).nodes.len();
-            chosen_gated.push(chosen);
-
-            // Deepest fort-free severing tile available on this map.
-            let mut best = 0usize;
-            for r in 0..base.rows() {
-                for c in 0..base.cols {
-                    let tile = base.get(r, c);
-                    if !LOCKABLE_TILES.contains(&tile) {
-                        continue;
-                    }
-                    let mut g2 = base.clone();
-                    g2.set(r, c, gap_tile_for(tile));
-                    let walk = walk_map(&g2, &built.pipe_pairs, start, wi);
-                    if walk.nodes.contains(&target) {
-                        continue;
-                    }
-                    if forts.iter().any(|f| !walk.nodes.contains(f)) {
-                        continue; // strands a fort — not usable headroom
-                    }
-                    best = best.max(open_nodes.len() - walk.nodes.len());
-                }
-            }
-            best_gated.push(best);
-            if chosen + 2 >= best {
-                at_headroom += 1;
-            }
-        }
-    }
-    eprintln!("\n=== goal_gate_depth_probe (300 seeds, {worlds_with_gate} gated worlds) ===");
-    eprint!("  gate->target Manhattan dist: ");
-    for (d, n) in dist_hist.iter().enumerate() {
-        if *n > 0 {
-            eprint!("{}{}={} ", if d == 11 { "≥" } else { "" }, d, n);
-        }
-    }
-    let mean = |v: &Vec<usize>| v.iter().sum::<usize>() as f64 / v.len().max(1) as f64;
-    eprintln!(
-        "\n  doorstep gates (dist <=2): {} ({:.1}%)",
-        dist_hist[..3].iter().sum::<usize>(),
-        100.0 * dist_hist[..3].iter().sum::<usize>() as f64 / worlds_with_gate as f64
-    );
-    eprintln!(
-        "  gated nodes: chosen mean {:.1} vs best-available (fort-free) mean {:.1}; chosen within 2 of best: {} ({:.1}%)",
-        mean(&chosen_gated),
-        mean(&best_gated),
-        at_headroom,
-        100.0 * at_headroom as f64 / worlds_with_gate as f64
-    );
-}
-
-/// Rule-aware dissect of SAS goal-open worlds: the rule-blind scan says the
-/// goal IS single-lock gateable (3-10 severing tiles), yet place_locks ends
-/// all-safe. For each severing tile, report WHICH lock rule refuses it:
-/// row-7/8 completion-bit exclusion, rule 1 (strands its own section's
-/// fort), rule 2 (strands an earlier section's fort), or GoalGate role
-/// rejection (strands any fort). Run with SAS=1.
-#[test]
-#[ignore]
-fn w7_sas_rule_dissect() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-    let mut dissected = 0usize;
-    for seed in 0..1000u64 {
-        if dissected >= 5 {
-            break;
-        }
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let result = build(
-            &rom,
-            &OverworldData { pickup: &pickup, catalog: &catalog },
-            &mut rng,
-            BuildFlags { shuffle_toad_houses: true, ..Default::default() },
-        );
-        for built in &result.worlds {
-            let open = world_topology(built).is_some_and(|t| t.fort_count >= 2 && t.depth == 0);
-            if !open {
-                continue;
-            }
-            dissected += 1;
-            let wi = built.world_idx;
-            let mut base = built.grid.clone();
-            stamp_slots(&mut base, &built.slots);
-            let start = rom_data::find_start(&base);
-            let target = find_target(&base, wi).unwrap();
-
-            // Fort positions by section, ordered.
-            let forts: Vec<(usize, (usize, usize))> = {
-                let mut f: Vec<(usize, (usize, usize))> = built
-                    .slots
-                    .iter()
-                    .filter(|s| s.kind == SlotKind::Fortress)
-                    .map(|s| (s.section, s.pos))
-                    .collect();
-                f.sort();
-                f
-            };
-            eprintln!("\nseed {seed} W{}: forts {:?}", wi + 1, forts);
-            for l in &built.locks {
-                eprintln!(
-                    "  placed lock: sec {} at {:?} safe={} blocks_target={}",
-                    l.fort_section, l.pos, l.secret_exit_safe, l.blocks_target
-                );
-            }
-
-            for r in 0..base.rows() {
-                for c in 0..base.cols {
-                    let tile = base.get(r, c);
-                    if !LOCKABLE_TILES.contains(&tile) {
-                        continue;
-                    }
-                    let mut g = base.clone();
-                    g.set(r, c, gap_tile_for(tile));
-                    let walk = walk_map(&g, &built.pipe_pairs, start, wi);
-                    if walk.nodes.contains(&target) {
-                        continue; // not goal-severing
-                    }
-                    // Row 7/8 completion-bit exclusion, as in place_locks.
-                    let row78 = (r == 7 || r == 8) && {
-                        let paired = if r == 7 { 8 } else { 7 };
-                        built.slots.iter().any(|s| {
-                            s.pos == (paired, c)
-                                && matches!(
-                                    s.kind,
-                                    SlotKind::Level
-                                        | SlotKind::Fortress
-                                        | SlotKind::Pipe
-                                        | SlotKind::BonusGame
-                                        | SlotKind::ToadHouse
-                                )
-                        })
-                    };
-                    let stranded: Vec<usize> = forts
-                        .iter()
-                        .filter(|(_, fp)| !walk.nodes.contains(fp))
-                        .map(|(sec, _)| *sec)
-                        .collect();
-                    eprintln!(
-                        "  severing ({r},{c}) tile {tile:02X}: row78_excluded={} strands_fort_sections={:?} gated_nodes={}",
-                        row78,
-                        stranded,
-                        {
-                            let open_walk = walk_map(&base, &built.pipe_pairs, start, wi);
-                            open_walk.nodes.len() - walk.nodes.len()
-                        },
-                    );
-                }
-            }
-        }
-    }
-}
-
-/// How NESTED are a world's locks? A chain-link lock is "nested" when closing
-/// it alone (others open) blocks the GOAL — i.e. its gated region contains
-/// the rest of the progression, not just the next fort's local pocket. The
-/// fort-skip machinery can only see nested locks (its splits are goal-anchored),
-/// so a world whose chain links gate local pockets is invisible to it.
-#[test]
-#[ignore]
-fn lock_nesting_probe() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-    // (nonfinal locks blocking goal, nonfinal total, final blocking, final total)
-    let mut per_world = [(0u64, 0u64, 0u64, 0u64); 8];
-    for seed in 0..300u64 {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let result = build(
-            &rom,
-            &OverworldData { pickup: &pickup, catalog: &catalog },
-            &mut rng,
-            BuildFlags { shuffle_toad_houses: true, ..Default::default() },
-        );
-        for built in &result.worlds {
-            if built.locks.len() < 2 {
-                continue;
-            }
-            let wi = built.world_idx;
-            let mut base = built.grid.clone();
-            stamp_slots(&mut base, &built.slots);
-            let Some(start) = rom_data::find_start(&base) else { continue };
-            let Some(target) = find_target(&base, wi) else { continue };
-            let all_secs: HashSet<usize> = built.locks.iter().map(|l| l.fort_section).collect();
-            let max_sec = *all_secs.iter().max().unwrap();
-            for l in &built.locks {
-                let mut g = base.clone();
-                for l2 in &built.locks {
-                    let t = if l2.fort_section == l.fort_section { l2.gap_tile } else { l2.replace_tile };
-                    g.set(l2.pos.0, l2.pos.1, t);
-                }
-                let blocks =
-                    !walk_map(&g, &built.pipe_pairs, Some(start), wi).nodes.contains(&target);
-                let s = &mut per_world[wi];
-                if l.fort_section == max_sec {
-                    s.3 += 1;
-                    if blocks { s.2 += 1; }
-                } else {
-                    s.1 += 1;
-                    if blocks { s.0 += 1; }
-                }
-            }
-        }
-    }
-    eprintln!("\n=== lock_nesting_probe (300 seeds, multi-lock worlds) ===");
-    eprintln!("world  nonfinal-locks-blocking-goal%   final-lock-blocking-goal%");
-    for (wi, &(nb, nt, fb, ft)) in per_world.iter().enumerate() {
-        if nt + ft == 0 { continue; }
-        eprintln!(
-            "  W{}   {:>6.1}% ({}/{})              {:>6.1}% ({}/{})",
-            wi + 1,
-            100.0 * nb as f64 / nt.max(1) as f64, nb, nt,
-            100.0 * fb as f64 / ft.max(1) as f64, fb, ft,
-        );
-    }
-}
-
-/// TEMP: dissect W2 fort-skip failures at placement-time fidelity — W2 has
-/// no connectivity pipes, so the placer's view is exactly pipes=[]. For the
-/// first N skip-allowed W2 worlds, print each lock's split and the candidate
-/// endpoint situation the spare-pipe pass actually faced.
-#[test]
-#[ignore]
-fn w2_skip_dissect() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-    let mut shown = 0;
-    for seed in 0..200u64 {
-        if shown >= 6 {
-            break;
-        }
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let result = build(
-            &rom,
-            &OverworldData { pickup: &pickup, catalog: &catalog },
-            &mut rng,
-            BuildFlags { shuffle_toad_houses: true, ..Default::default() },
-        );
-        let built = &result.worlds[1]; // W2
-        if built.locks.len() < 2 {
-            continue;
-        }
-        shown += 1;
-        let mut base = built.grid.clone();
-        stamp_slots(&mut base, &built.slots);
-        let start = rom_data::find_start(&base).unwrap();
-        let target = find_target(&base, 1).unwrap();
-        let reserved: HashSet<Pos> =
-            rom_data::read_hb_sprite_positions(&rom, 1).into_iter().collect();
-        let lock_row78: HashSet<Pos> = built
-            .locks
-            .iter()
-            .filter_map(|l| match l.pos.0 {
-                7 => Some((8, l.pos.1)),
-                8 => Some((7, l.pos.1)),
-                _ => None,
-            })
-            .collect();
-        let endpoints: Vec<Pos> = built
-            .slots
-            .iter()
-            .filter(|s| {
-                matches!(s.kind, SlotKind::HammerBro | SlotKind::ToadHouse | SlotKind::BonusGame)
-                    && !reserved.contains(&s.pos)
-                    && !lock_row78.contains(&s.pos)
-            })
-            .map(|s| s.pos)
-            .collect();
-        eprintln!(
-            "\nseed {seed} W2: pipes {}, forts {}, HB-ish endpoints {} (reserved {})",
-            built.pipe_pairs.len(),
-            built.slots.iter().filter(|s| s.kind == SlotKind::Fortress).count(),
-            endpoints.len(),
-            reserved.len(),
-        );
-        let all_secs: HashSet<usize> = built.locks.iter().map(|l| l.fort_section).collect();
-        for l in &built.locks {
-            let mut g = base.clone();
-            for l2 in &built.locks {
-                let t = if l2.fort_section == l.fort_section { l2.gap_tile } else { l2.replace_tile };
-                g.set(l2.pos.0, l2.pos.1, t);
-            }
-            // Placement-time view: no pipes (W2 has no connectivity pipes).
-            let cs = walk_map(&g, &[], Some(start), 1).nodes;
-            let blocks = !cs.contains(&target);
-            if !blocks {
-                eprintln!(
-                    "  lock sec {} at {:?}: does NOT block goal (others open) -> invisible to skip eval",
-                    l.fort_section, l.pos
-                );
-                continue;
-            }
-            let ct = walk_map(&g, &[], Some(target), 1).nodes;
-            let ea = endpoints.iter().filter(|p| cs.contains(p)).count();
-            let eb = endpoints.iter().filter(|p| ct.contains(p)).count();
-            // Unclassified nodes: reachable from neither (stranded by this lock).
-            let neither = endpoints.iter().filter(|p| !cs.contains(p) && !ct.contains(p)).count();
-            eprintln!(
-                "  lock sec {} at {:?}: BLOCKS goal; start-side {} nodes/{} endpts, goal-side {} nodes/{} endpts, neither {}",
-                l.fort_section, l.pos, cs.len(), ea, ct.len(), eb, neither
-            );
-        }
-        // All-closed split (the goal-open rejection state).
-        let mut g = base.clone();
-        for l2 in &built.locks {
-            g.set(l2.pos.0, l2.pos.1, l2.gap_tile);
-        }
-        let cs = walk_map(&g, &[], Some(start), 1).nodes;
-        let ct = walk_map(&g, &[], Some(target), 1).nodes;
-        let ea = endpoints.iter().filter(|p| cs.contains(p)).count();
-        let eb = endpoints.iter().filter(|p| ct.contains(p)).count();
-        eprintln!(
-            "  ALL-CLOSED: start-side {} endpts, goal-side {} endpts (a straddle here = goal-open reject)",
-            ea, eb
-        );
-        // Where did the actual spare pipe land relative to each gating lock?
-        for &(a, b) in &built.pipe_pairs {
-            for l in &built.locks {
-                let mut g = base.clone();
-                for l2 in &built.locks {
-                    let t = if l2.fort_section == l.fort_section { l2.gap_tile } else { l2.replace_tile };
-                    g.set(l2.pos.0, l2.pos.1, t);
-                }
-                let cs = walk_map(&g, &[], Some(start), 1).nodes;
-                if cs.contains(&target) {
-                    continue;
-                }
-                let ct = walk_map(&g, &[], Some(target), 1).nodes;
-                let side = |p: Pos| {
-                    if cs.contains(&p) { "start" } else if ct.contains(&p) { "goal" } else { "mid" }
-                };
-                eprintln!(
-                    "  placed pipe {:?}<->{:?} vs lock sec {}: {} <-> {}",
-                    a, b, l.fort_section, side(a), side(b)
-                );
-            }
-        }
-        let _ = all_secs;
-    }
-}
-
-/// Per-world pipe budget composition: how many of each world's pairs are
-/// island bridges (an endpoint unreachable by walk alone, locks open) vs
-/// spares (both endpoints on the walk-connected mainland). Bridges are
-/// placed by the connectivity phase; only spares can become fort-skips.
-#[test]
-#[ignore]
-fn pipe_budget_probe() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-    let mut per_world = [(0u64, 0u64, 0u64); 8]; // (pairs, bridges, worlds)
-    for seed in 0..300u64 {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let result = build(
-            &rom,
-            &OverworldData { pickup: &pickup, catalog: &catalog },
-            &mut rng,
-            BuildFlags { shuffle_toad_houses: true, ..Default::default() },
-        );
-        for built in &result.worlds {
-            let wi = built.world_idx;
-            let mut base = built.grid.clone();
-            stamp_slots(&mut base, &built.slots);
-            let Some(start) = rom_data::find_start(&base) else { continue };
-            let open_walk = walk_map(&base, &[], Some(start), wi).nodes;
-            let bridges = built
-                .pipe_pairs
-                .iter()
-                .filter(|&&(a, b)| !open_walk.contains(&a) || !open_walk.contains(&b))
-                .count();
-            let s = &mut per_world[wi];
-            s.0 += built.pipe_pairs.len() as u64;
-            s.1 += bridges as u64;
-            s.2 += 1;
-        }
-    }
-    eprintln!("\n=== pipe_budget_probe (300 seeds) ===");
-    eprintln!("world  avg pairs  avg bridges  avg spares");
-    for (wi, &(pairs, bridges, n)) in per_world.iter().enumerate() {
-        if n == 0 { continue; }
-        let n = n as f64;
-        eprintln!(
-            "  W{}     {:>5.2}      {:>5.2}       {:>5.2}",
-            wi + 1,
-            pairs as f64 / n,
-            bridges as f64 / n,
-            (pairs - bridges) as f64 / n,
-        );
-    }
-}
-
-
-
-/// Player-model fort taxonomy over many seeds. Classifies every fortress that
-/// owns a lock into: mandatory (goal unreachable without it), shortcut (optional
-/// but beating it cuts >=2 levels off the route, counting the fort itself),
-/// decoy (optional, no worthwhile shortcut, but accessible at start so it reads
-/// as a real choice), or inert (optional, off the beaten path). Also reports how
-/// often a world is a "choice-fork" (one required fort plus >=1 accessible decoy)
-/// and the distribution of levels a shortcut actually saves. Run with:
-///   cargo test --lib fort_taxonomy -- --ignored --nocapture
-#[test]
-#[ignore]
-fn fort_taxonomy() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-    const SEEDS: u64 = 1000;
-
-    let mut n_mandatory = 0u64;
-    let mut n_shortcut = 0u64;
-    let mut n_decoy = 0u64;
-    let mut n_inert = 0u64;
-    let mut n_nolock = 0u64; // forts with no lock at all
-    let mut saved_hist = [0u64; 8]; // levels saved by a shortcut fort (capped 7)
-
-    // World-level buckets.
-    let mut w_total = 0u64;
-    let mut w_choice_fork = 0u64; // 1 required + >=1 accessible decoy
-    let mut w_has_shortcut = 0u64;
-    // Cross-tab of (chain depth to goal) x (has a fork), 2+-fort worlds only.
-    let mut b_pure_chain = 0u64; // depth >=2, no fork  (the "grind" we dislike)
-    let mut b_chain_fork = 0u64; // depth >=2, has fork (W8 chain-2+fork-2)
-    let mut b_fork_shallow = 0u64; // depth <2, has fork (fork at/near start)
-    let mut b_linear = 0u64; // depth <2, no fork
-    let mut w_multi = 0u64;
-
-    let results = par_seeds(SEEDS, |seed| {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        build(
-            &rom,
-            &OverworldData { pickup: &pickup, catalog: &catalog },
-            &mut rng,
-            BuildFlags { shuffle_toad_houses: true, ..Default::default() },
-        )
-    });
-    for result in &results {
-        for built in &result.worlds {
-            let wi = built.world_idx;
-            let mut base = built.grid.clone();
-            stamp_slots(&mut base, &built.slots);
-            let Some(start) = rom_data::find_start(&base) else { continue };
-            let Some(target) = find_target(&base, wi) else { continue };
-
-            let forts: Vec<(usize, Pos)> = built
-                .slots
-                .iter()
-                .filter(|s| s.kind == SlotKind::Fortress)
-                .map(|s| (s.section, s.pos))
-                .collect();
-            if forts.is_empty() {
-                continue;
-            }
-            let blocking: HashSet<Pos> = built
-                .slots
-                .iter()
-                .filter(|s| matches!(s.kind, SlotKind::Level | SlotKind::Fortress))
-                .map(|s| s.pos)
-                .collect();
-            let all_sections: HashSet<usize> =
-                built.locks.iter().map(|l| l.fort_section).collect();
-
-            let grid_with = |opened: &HashSet<usize>| -> Grid {
-                let mut g = base.clone();
-                for l in &built.locks {
-                    if opened.contains(&l.fort_section) {
-                        g.set(l.pos.0, l.pos.1, l.replace_tile);
-                    } else {
-                        g.set(l.pos.0, l.pos.1, l.gap_tile);
-                    }
-                }
-                g
-            };
-
-            // Progression simulation: the round at which each fort first
-            // becomes reachable (beat every reachable fort each round). Forts
-            // sharing a round are co-accessible — a fork. Detects forks at any
-            // depth, including a fork behind a chain (W8's chain-2 + fork-2).
-            let mut unlock_round: HashMap<usize, usize> = HashMap::new();
-            let mut goal_round: Option<usize> = None;
-            {
-                let mut beaten: HashSet<usize> = HashSet::new();
-                let mut round = 0usize;
-                loop {
-                    let g = grid_with(&beaten);
-                    let reach = walk_map(&g, &built.pipe_pairs, Some(start), wi);
-                    if goal_round.is_none() && reach.nodes.contains(&target) {
-                        goal_round = Some(round);
-                    }
-                    let newly: Vec<usize> = forts
-                        .iter()
-                        .filter(|(s, p)| !unlock_round.contains_key(s) && reach.nodes.contains(p))
-                        .map(|(s, _)| *s)
-                        .collect();
-                    if newly.is_empty() {
-                        break;
-                    }
-                    for &s in &newly {
-                        unlock_round.insert(s, round);
-                        beaten.insert(s);
-                    }
-                    round += 1;
-                    if round > 30 {
-                        break;
-                    }
-                }
-            }
-            // Cohort size = how many forts share an unlock round.
-            let mut round_counts: HashMap<usize, usize> = HashMap::new();
-            for r in unlock_round.values() {
-                *round_counts.entry(*r).or_insert(0) += 1;
-            }
-
-            let mut world_decoys = 0u64;
-            let mut world_shortcuts = 0u64;
-
-            for &(sec, fpos) in &forts {
-                if !all_sections.contains(&sec) {
-                    n_nolock += 1;
-                    continue;
-                }
-                let cohort = unlock_round
-                    .get(&sec)
-                    .and_then(|r| round_counts.get(r))
-                    .copied()
-                    .unwrap_or(1);
-
-                // All other locks open, this one's state varies.
-                let mut others_open = all_sections.clone();
-                others_open.remove(&sec);
-                let g_closed = grid_with(&others_open); // this fort NOT beaten
-                let g_open = grid_with(&all_sections); // this fort beaten too
-
-                // Mandatory: goal unreachable while this lock stays closed.
-                let reach_closed = walk_map(&g_closed, &built.pipe_pairs, Some(start), wi);
-                if !reach_closed.nodes.contains(&target) {
-                    n_mandatory += 1;
-                    continue;
-                }
-
-                // Optional fort. Compute shortcut value in levels.
-                let from_start_closed = least_levels_from(&g_closed, &built.pipe_pairs, start, wi, &blocking);
-                let from_fort_open = least_levels_from(&g_open, &built.pipe_pairs, fpos, wi, &blocking);
-                let l_without = from_start_closed.get(&target).copied();
-                // reach+beat F = dist to F (F is a fort, so its weight of 1 is
-                // already the "beat it" cost), then F -> goal via opened door.
-                let l_with = match (from_start_closed.get(&fpos), from_fort_open.get(&target)) {
-                    (Some(&to_f), Some(&f_to_goal)) => Some(to_f + f_to_goal),
-                    _ => None,
-                };
-
-                let saved = match (l_without, l_with) {
-                    (Some(wo), Some(wi_)) => wo as i64 - wi_ as i64,
-                    _ => i64::MIN,
-                };
-
-                if saved >= 2 {
-                    n_shortcut += 1;
-                    world_shortcuts += 1;
-                    saved_hist[(saved as usize).min(7)] += 1;
-                } else if cohort >= 2 {
-                    // Optional fort co-accessible with a sibling = a decoy in a fork.
-                    n_decoy += 1;
-                    world_decoys += 1;
-                } else {
-                    n_inert += 1;
-                }
-            }
-
-            w_total += 1;
-            if world_decoys >= 1 {
-                w_choice_fork += 1;
-            }
-            if world_shortcuts >= 1 {
-                w_has_shortcut += 1;
-            }
-
-            // Cross-tab: pure grind chain vs chain-then-fork, 2+-fort worlds.
-            if forts.len() >= 2 {
-                w_multi += 1;
-                let deep = goal_round.unwrap_or(0) >= 2;
-                let has_fork = world_decoys >= 1;
-                match (deep, has_fork) {
-                    (true, false) => b_pure_chain += 1,
-                    (true, true) => b_chain_fork += 1,
-                    (false, true) => b_fork_shallow += 1,
-                    (false, false) => b_linear += 1,
-                }
-            }
-        }
-    }
-
-    let fort_total = n_mandatory + n_shortcut + n_decoy + n_inert;
-    eprintln!("\n=== Fort taxonomy ({fort_total} forts w/ a lock; {n_nolock} forts had no lock), {SEEDS} seeds ===");
-    let fp = |c: u64| 100.0 * c as f64 / fort_total as f64;
-    eprintln!("  mandatory (goal needs it)      : {n_mandatory:6} ({:5.1}%)", fp(n_mandatory));
-    eprintln!("  shortcut  (optional, saves >=2): {n_shortcut:6} ({:5.1}%)", fp(n_shortcut));
-    eprintln!("  decoy     (optional, accessible, no shortcut): {n_decoy:6} ({:5.1}%)", fp(n_decoy));
-    eprintln!("  inert     (optional, off-path) : {n_inert:6} ({:5.1}%)", fp(n_inert));
-
-    eprintln!("\n=== Levels saved by shortcut forts ({n_shortcut} forts) ===");
-    let stot: u64 = saved_hist.iter().sum();
-    for (s, &c) in saved_hist.iter().enumerate() {
-        if c == 0 {
-            continue;
-        }
-        let label = if s == 7 { "7+".to_string() } else { s.to_string() };
-        eprintln!("  saves {label}: {c:6} ({:5.1}%)", 100.0 * c as f64 / stot as f64);
-    }
-
-    eprintln!("\n=== World-level ({w_total} worlds) ===");
-    eprintln!("  choice-fork (>=1 co-accessible optional decoy, any depth): {w_choice_fork:6} ({:5.1}%)", 100.0 * w_choice_fork as f64 / w_total as f64);
-    eprintln!("  has >=1 shortcut fort: {w_has_shortcut:6} ({:5.1}%)", 100.0 * w_has_shortcut as f64 / w_total as f64);
-
-    eprintln!("\n=== Chain x Fork cross-tab ({w_multi} worlds w/ 2+ forts) ===");
-    let mp = |c: u64| 100.0 * c as f64 / w_multi as f64;
-    eprintln!("  PURE CHAIN  (depth>=2, no fork — the grind): {b_pure_chain:6} ({:5.1}%)", mp(b_pure_chain));
-    eprintln!("  chain+fork  (depth>=2, has fork — e.g. W8): {b_chain_fork:6} ({:5.1}%)", mp(b_chain_fork));
-    eprintln!("  fork shallow(depth<2,  has fork)          : {b_fork_shallow:6} ({:5.1}%)", mp(b_fork_shallow));
-    eprintln!("  linear      (depth<2,  no fork)           : {b_linear:6} ({:5.1}%)", mp(b_linear));
-}
 
 /// Baseline metrics for the lock/fort progression-topology work. Emits, over
 /// many seeds:
@@ -1958,110 +922,6 @@ fn progression_metrics() {
     }
 }
 
-/// Diagnostic: in worlds with 2+ fortresses, how many forts are reachable from
-/// the world start *before completing any fortress* (all locks closed)? A later
-/// fort can be gated behind an earlier fort's lock, so not all are guaranteed
-/// open at game start. Run with:
-///   cargo test --lib forts_accessible_at_start -- --ignored --nocapture
-#[test]
-#[ignore]
-fn forts_accessible_at_start() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-
-    const SEEDS: u64 = 1000;
-    // For a world with N forts, tally how many are accessible at start.
-    // by_fortcount[N][accessible] — N in 2..=4, accessible in 0..=N.
-    let mut by_fortcount = [[0u32; 5]; 5];
-    // Totals across every world that has >= 2 forts.
-    let mut worlds_2plus = 0u64;
-    let mut total_forts_2plus = 0u64;
-    let mut total_accessible_2plus = 0u64;
-    let mut worlds_all_open = 0u64; // every fort reachable at start
-    let mut worlds_some_locked = 0u64; // at least one fort gated behind another
-
-    for seed in 0..SEEDS {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let result = build(
-            &rom,
-            &OverworldData { pickup: &pickup, catalog: &catalog },
-            &mut rng,
-            BuildFlags { shuffle_toad_houses: true, ..Default::default() },
-        );
-
-        for built in &result.worlds {
-            let wi = built.world_idx;
-            let fort_positions: Vec<Pos> = built
-                .slots
-                .iter()
-                .filter(|s| s.kind == SlotKind::Fortress)
-                .map(|s| s.pos)
-                .collect();
-            let fort_count = fort_positions.len();
-            if fort_count < 2 {
-                continue;
-            }
-
-            // Game-start grid: slots stamped, then ALL locks closed (gapped).
-            let mut grid = built.grid.clone();
-            stamp_slots(&mut grid, &built.slots);
-            for lock in &built.locks {
-                grid.set(lock.pos.0, lock.pos.1, lock.gap_tile);
-            }
-
-            let start = rom_data::find_start(&grid);
-            let walk = walk_map(&grid, &built.pipe_pairs, start, wi);
-            let accessible = fort_positions.iter().filter(|p| walk.nodes.contains(*p)).count();
-
-            by_fortcount[fort_count][accessible] += 1;
-            worlds_2plus += 1;
-            total_forts_2plus += fort_count as u64;
-            total_accessible_2plus += accessible as u64;
-            if accessible == fort_count {
-                worlds_all_open += 1;
-            } else {
-                worlds_some_locked += 1;
-            }
-        }
-    }
-
-    eprintln!("\nForts accessible at start (all locks closed), {SEEDS} seeds:");
-    eprintln!("  For each world with N forts, how many are reachable before beating any fort:");
-    for (n, row) in by_fortcount.iter().enumerate().take(5).skip(2) {
-        let total: u32 = row.iter().sum();
-        if total == 0 {
-            continue;
-        }
-        eprint!("  {n}-fort worlds ({total:5} instances): ");
-        for (a, &c) in row.iter().enumerate().take(n + 1) {
-            eprint!("{a}→{:5.1}%  ", 100.0 * c as f64 / total as f64);
-        }
-        let mean: f64 = row.iter().take(n + 1).enumerate().map(|(a, &c)| a as f64 * c as f64).sum::<f64>() / total as f64;
-        eprintln!(" (mean {mean:.2})");
-    }
-
-    eprintln!("\nAcross all {worlds_2plus} world-instances with 2+ forts:");
-    eprintln!(
-        "  {} forts total, {} accessible at start = {:.1}% openable without beating a fort",
-        total_forts_2plus,
-        total_accessible_2plus,
-        100.0 * total_accessible_2plus as f64 / total_forts_2plus as f64,
-    );
-    eprintln!(
-        "  all forts open at start: {} worlds ({:.1}%)",
-        worlds_all_open,
-        100.0 * worlds_all_open as f64 / worlds_2plus as f64,
-    );
-    eprintln!(
-        "  >=1 fort gated behind another: {} worlds ({:.1}%)",
-        worlds_some_locked,
-        100.0 * worlds_some_locked as f64 / worlds_2plus as f64,
-    );
-}
-
 #[test]
 #[ignore]
 fn test_print_build() {
@@ -2282,586 +1142,6 @@ fn test_w6_slot_distribution() {
     }
 }
 
-#[test]
-#[ignore]
-fn test_dump_w7_blank_vs_bfs() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-    let wi = 6; // W7
-
-    let cw = &pickup.worlds[wi];
-    eprintln!("\n=== W7 Pickup: {} pool entries ===", cw.pool_indices.len());
-
-    let fixed = fixed_positions_for_world(&rom, &catalog, wi, true, false);
-    eprintln!("Fixed positions: {} {:?}", fixed.len(), fixed);
-
-    let blank_positions = find_blank_slots(&cw.grid, &fixed);
-    eprintln!("Blank tiles on grid: {}", blank_positions.len());
-
-    // Run the actual build for several seeds and check coverage
-    for seed in 0..5u64 {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let result = build(&rom, &OverworldData { pickup: &pickup, catalog: &catalog }, &mut rng, BuildFlags { shuffle_toad_houses: true, ..Default::default() });
-        let built = &result.worlds[wi];
-
-        // All positions that got a slot assignment
-        let slot_positions: HashSet<(usize, usize)> = built.slots.iter().map(|s| s.pos).collect();
-        // Add pipe positions
-        let pipe_positions: HashSet<(usize, usize)> = built.pipe_pairs.iter()
-            .flat_map(|&(a, b)| vec![a, b]).collect();
-        let all_assigned: HashSet<(usize, usize)> = slot_positions.union(&pipe_positions).copied().collect();
-
-        // Blank tiles with no assignment
-        let uncovered: Vec<(usize, usize)> = blank_positions.iter()
-            .filter(|p| !all_assigned.contains(p))
-            .copied()
-            .collect();
-
-        let total_slots = built.slots.len() + pipe_positions.len();
-        eprintln!("\n--- Seed {seed} ---");
-        eprintln!("  Slots: {} (L={}, F={}, P={}, HB={})",
-            total_slots,
-            built.slots.iter().filter(|s| s.kind == SlotKind::Level).count(),
-            built.slots.iter().filter(|s| s.kind == SlotKind::Fortress).count(),
-            pipe_positions.len(),
-            built.slots.iter().filter(|s| s.kind == SlotKind::HammerBro).count(),
-        );
-        eprintln!("  Pool entries (ptr slots): {}", cw.pool_indices.len());
-        eprintln!("  max_non_pipe_slots: {}", cw.pool_indices.len() - VANILLA_PIPE_PAIRS[wi] * 2);
-        eprintln!("  Blanks on grid: {}", blank_positions.len());
-        eprintln!("  Assigned positions: {}", all_assigned.len());
-        eprintln!("  Uncovered blanks: {}", uncovered.len());
-
-        if !uncovered.is_empty() {
-            for (r, c) in &uncovered {
-                eprintln!("    UNCOVERED: ({},{}) tile=${:02X}", r, c, cw.grid.get(*r, *c));
-                // Check if BFS can reach it with the placed pipes
-                let bfs_all = bfs_ordered(&built.grid, &built.pipe_pairs, rom_data::find_start(&built.grid), built.world_idx);
-                let bfs_set: HashSet<(usize, usize)> = bfs_all.iter().map(|&(p, _)| p).collect();
-                eprintln!("      BFS reachable: {}", bfs_set.contains(&(*r, *c)));
-            }
-        }
-
-        // Check for assignments NOT on blank tiles (double-covering or wrong pos)
-        let non_blank_assignments: Vec<_> = all_assigned.iter()
-            .filter(|p| !blank_positions.contains(p) && !pipe_positions.contains(p))
-            .collect();
-        if !non_blank_assignments.is_empty() {
-            eprintln!("  Assignments on non-blank tiles:");
-            for &&(r, c) in &non_blank_assignments {
-                eprintln!("    ({},{}) tile=${:02X}", r, c, cw.grid.get(r, c));
-            }
-        }
-    }
-}
-
-#[test]
-#[ignore]
-fn test_lock_scoring_detail() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => {
-            eprintln!("ROM not found, skipping");
-            return;
-        }
-    };
-    let (catalog, pickup) = build_catalog_pickup(&rom, 0);
-
-    for seed in [42u64, 123, 999] {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let result = build(&rom, &OverworldData { pickup: &pickup, catalog: &catalog }, &mut rng, BuildFlags { shuffle_toad_houses: true, ..Default::default() });
-
-        eprintln!("\n{}", "=".repeat(60));
-        eprintln!("=== Seed {seed} ===");
-
-        for built in &result.worlds {
-            let wi = built.world_idx;
-            let target_pos = find_target(&built.grid, wi);
-            let start_pos = rom_data::find_start(&built.grid);
-
-            let forts: Vec<_> = built.slots.iter()
-                .filter(|s| s.kind == SlotKind::Fortress)
-                .collect();
-            let levels: Vec<_> = built.slots.iter()
-                .filter(|s| s.kind == SlotKind::Level)
-                .collect();
-
-            eprintln!("\n  W{}: {} forts, {} levels, {} pipes, {} locks, target={:?}",
-                wi + 1, forts.len(), levels.len(), built.pipe_pairs.len(),
-                built.locks.len(), target_pos);
-
-            // Build stamped grid (no locks), same as production
-            let mut base_grid = built.grid.clone();
-            stamp_slots(&mut base_grid, &built.slots);
-
-            for lock in &built.locks {
-                // Open grid: no locks
-                let walk_open = walk_map(&base_grid, &built.pipe_pairs, start_pos, built.world_idx);
-
-                // Locked grid: this lock closed
-                let mut locked_grid = base_grid.clone();
-                locked_grid.set(lock.pos.0, lock.pos.1, lock.gap_tile);
-                let walk_locked = walk_map(&locked_grid, &built.pipe_pairs, start_pos, built.world_idx);
-
-                let gated_count = walk_open.nodes.len() as i32 - walk_locked.nodes.len() as i32;
-
-                // What specifically gets gated?
-                let gated_forts: Vec<_> = forts.iter()
-                    .filter(|f| walk_open.nodes.contains(&f.pos) && !walk_locked.nodes.contains(&f.pos))
-                    .collect();
-                let gated_levels: Vec<_> = levels.iter()
-                    .filter(|l| walk_open.nodes.contains(&l.pos) && !walk_locked.nodes.contains(&l.pos))
-                    .collect();
-                let gates_target = target_pos
-                    .map(|tp| walk_open.nodes.contains(&tp) && !walk_locked.nodes.contains(&tp))
-                    .unwrap_or(false);
-
-                // BFS distance from lock to target (via adjacent nodes)
-                let target_dist = if let Some(tp) = target_pos {
-                    let walk_from_target = walk_map(&base_grid, &built.pipe_pairs, Some(tp), built.world_idx);
-                    let (lr, lc) = lock.pos;
-                    [(-1i16, 0i16), (1, 0), (0, -1), (0, 1)].iter()
-                        .filter_map(|&(dr, dc)| {
-                            let nr = lr as i16 + dr;
-                            let nc = lc as i16 + dc;
-                            if nr < 0 || nc < 0 { return None; }
-                            walk_from_target.distances.get(&(nr as usize, nc as usize)).copied()
-                        })
-                        .min()
-                } else {
-                    None
-                };
-
-                eprintln!("    Lock ({:2},{:2}) sect={} safe={:<5} gated={:<3} dist_to_target={:<4} gates: {} forts, {} levels{}",
-                    lock.pos.0, lock.pos.1,
-                    lock.fort_section,
-                    lock.secret_exit_safe,
-                    gated_count,
-                    target_dist.map(|d| d.to_string()).unwrap_or("-".into()),
-                    gated_forts.len(),
-                    gated_levels.len(),
-                    if gates_target { ", TARGET" } else { "" },
-                );
-            }
-        }
-    }
-}
-
-/// Dump all lock candidates and their scores for a specific world.
-/// Usage: change seed/target_wi below, then run with --nocapture.
-#[test]
-#[ignore]
-fn test_lock_candidates_dump() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => { eprintln!("ROM not found"); return; }
-    };
-    let seed = 42u64;
-    let target_wi = 6; // 0-indexed: W7 = 6
-
-    let (catalog, pickup) = build_catalog_pickup(&rom, seed);
-    let mut rng = ChaCha8Rng::seed_from_u64(seed);
-    let result = build(&rom, &OverworldData { pickup: &pickup, catalog: &catalog }, &mut rng, BuildFlags { shuffle_toad_houses: true, ..Default::default() });
-    let built = &result.worlds[target_wi];
-
-    let start_pos = rom_data::find_start(&built.grid);
-    let target_pos = find_target(&built.grid, target_wi);
-
-    // Build base grid with slots stamped (no locks), same as production
-    let mut base_grid = built.grid.clone();
-    stamp_slots(&mut base_grid, &built.slots);
-
-    eprintln!("\n=== Seed {seed}, W{} — Lock Candidate Dump ===", target_wi + 1);
-    eprintln!("Target: {:?}, Start: {:?}", target_pos, start_pos);
-    eprintln!("Forts: {}, Sections: {}", result.fort_counts[target_wi], built.section_count);
-
-    // For each section, enumerate all lockable tiles and score them
-    for section_idx in 0..built.section_count {
-        let fort_pos = match built.slots.iter()
-            .find(|s| s.section == section_idx && s.kind == SlotKind::Fortress)
-        {
-            Some(s) => s.pos,
-            None => continue,
-        };
-
-        eprintln!("\n  Section {section_idx} (fort at {:?}):", fort_pos);
-
-        // Open grid for this section: earlier locks open, no current lock
-        let walk_open = walk_map(&base_grid, &built.pipe_pairs, start_pos, built.world_idx);
-        let open_node_count = walk_open.nodes.len();
-
-        // Find all lockable path tiles
-        // (pos, gated, safe, score, blocks_later_fort, blocks_target)
-        type LockDebugCandidate = (Pos, i32, bool, i32, bool, bool);
-        let mut candidates: Vec<LockDebugCandidate> = Vec::new();
-
-        for r in 0..base_grid.rows() {
-            for c in 0..base_grid.cols {
-                let tile = base_grid.get(r, c);
-                if !LOCKABLE_TILES.contains(&tile) { continue; }
-
-                let gap = gap_tile_for(tile);
-                let mut test_grid = base_grid.clone();
-                test_grid.set(r, c, gap);
-                let walk = walk_map(&test_grid, &built.pipe_pairs, start_pos, built.world_idx);
-
-                // Hard rule: fort must be reachable
-                if !walk.nodes.contains(&fort_pos) { continue; }
-
-                let gated = open_node_count as i32 - walk.nodes.len() as i32;
-                let target_reachable = target_pos
-                    .map(|tp| walk.nodes.contains(&tp))
-                    .unwrap_or(true);
-                let safe = target_reachable && built.slots.iter().all(|s| {
-                    s.kind != SlotKind::Fortress || walk.nodes.contains(&s.pos)
-                });
-                let blocks_later_fort = built.slots.iter().any(|s| {
-                    s.kind == SlotKind::Fortress
-                        && s.section > section_idx
-                        && !walk.nodes.contains(&s.pos)
-                });
-                let mut score = gated;
-                if blocks_later_fort { score += 100; }
-
-                candidates.push(((r, c), gated, safe, score, blocks_later_fort, !target_reachable));
-            }
-        }
-
-        // Sort by score descending
-        candidates.sort_by_key(|c| std::cmp::Reverse(c.3));
-
-        let chosen = built.locks.iter().find(|l| l.fort_section == section_idx);
-        eprintln!("    {} candidates pass hard rules, chosen={:?}",
-            candidates.len(),
-            chosen.map(|l| l.pos));
-
-        for (pos, gated, safe, score, blf, bt) in &candidates {
-            let marker = if chosen.map(|l| l.pos == *pos).unwrap_or(false) { " <-- CHOSEN" } else { "" };
-            eprintln!("    ({:2},{:2}) gated={:<3} score={:<4} safe={:<5} blk_fort={:<5} blk_target={}{marker}",
-                pos.0, pos.1, gated, score, safe, blf, bt);
-        }
-    }
-}
-
-#[test]
-#[ignore]
-fn test_lock_airship_distance() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => {
-            eprintln!("ROM not found, skipping");
-            return;
-        }
-    };
-    let rom = apply_qol_for_overworld(&rom);
-
-    let seeds: u64 = std::env::var("LOCK_SEEDS")
-        .ok()
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(100);
-    // BFS distance histogram: index = distance, value = count
-    let mut histogram = [0u32; 30];
-    let mut total_locks = 0u32;
-    let mut no_target_locks = 0u32;
-    // Per-world stats: (sum_of_distances, count)
-    let mut per_world: [(u64, u32); 8] = [(0, 0); 8];
-    // Track locks at distance <= 2 per seed for flagging
-    let mut close_lock_seeds = 0u32;
-    // Inter-lock Manhattan distance (only for worlds with 2+ locks)
-    let mut inter_hist = [0u32; 40];
-    let mut total_pairs = 0u32;
-    let mut per_world_pairs: [(u64, u32); 8] = [(0, 0); 8];
-    let mut close_pair_seeds = 0u32;
-
-    for seed in 0..seeds {
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let (catalog, pickup) = build_catalog_pickup(&rom, seed);
-        let result = build(&rom, &OverworldData { pickup: &pickup, catalog: &catalog }, &mut rng, BuildFlags { shuffle_toad_houses: true, ..Default::default() });
-        let mut seed_has_close = false;
-        let mut seed_has_close_pair = false;
-
-        for built in &result.worlds {
-            let wi = built.world_idx;
-            let target_pos = find_target(&built.grid, wi);
-
-            // Inter-lock Manhattan distance (works regardless of target).
-            if built.locks.len() >= 2 {
-                for i in 0..built.locks.len() {
-                    for j in (i + 1)..built.locks.len() {
-                        let (ar, ac) = built.locks[i].pos;
-                        let (br, bc) = built.locks[j].pos;
-                        let d = ar.abs_diff(br) + ac.abs_diff(bc);
-                        let idx = d.min(inter_hist.len() - 1);
-                        inter_hist[idx] += 1;
-                        total_pairs += 1;
-                        per_world_pairs[wi].0 += d as u64;
-                        per_world_pairs[wi].1 += 1;
-                        if d <= 3 {
-                            seed_has_close_pair = true;
-                        }
-                    }
-                }
-            }
-
-            if target_pos.is_none() {
-                no_target_locks += built.locks.len() as u32;
-                continue;
-            }
-            let tp = target_pos.unwrap();
-
-            // Build a fully-stamped grid with all locks open so BFS
-            // reflects the walkable map. walk_map uses node-to-node
-            // hops (nodes are 2 tiles apart), so lock path tiles
-            // won't appear in distances. Instead, BFS from the target
-            // and measure to the node(s) adjacent to each lock.
-            let mut stamped = built.grid.clone();
-            stamp_slots(&mut stamped, &built.slots);
-
-            // BFS from target — distances to every reachable node
-            let walk_from_target = walk_map(&stamped, &built.pipe_pairs, Some(tp), built.world_idx);
-
-            for lock in &built.locks {
-                total_locks += 1;
-
-                // Lock is on a path tile between two nodes. Find the
-                // closest adjacent node (in BFS hops from target).
-                let (lr, lc) = lock.pos;
-                let adjacent_nodes: Vec<(usize, usize)> = [(-1i16, 0i16), (1, 0), (0, -1), (0, 1)]
-                    .iter()
-                    .filter_map(|&(dr, dc)| {
-                        let nr = lr as i16 + dr;
-                        let nc = lc as i16 + dc;
-                        if nr < 0 || nr >= stamped.rows() as i16 || nc < 0 || nc >= stamped.cols as i16 {
-                            return None;
-                        }
-                        let pos = (nr as usize, nc as usize);
-                        // Only count positions that are actual BFS nodes
-                        if walk_from_target.distances.contains_key(&pos) {
-                            Some(pos)
-                        } else {
-                            None
-                        }
-                    })
-                    .collect();
-
-                // Use the minimum distance among adjacent nodes
-                // (the side closer to the target).
-                let min_dist = adjacent_nodes.iter()
-                    .filter_map(|pos| walk_from_target.distances.get(pos))
-                    .min()
-                    .copied();
-
-                if let Some(dist) = min_dist {
-                    let idx = dist.min(histogram.len() - 1);
-                    histogram[idx] += 1;
-                    per_world[wi].0 += dist as u64;
-                    per_world[wi].1 += 1;
-
-                    if dist <= 2 {
-                        seed_has_close = true;
-                    }
-                } else {
-                    no_target_locks += 1;
-                }
-            }
-        }
-        if seed_has_close {
-            close_lock_seeds += 1;
-        }
-        if seed_has_close_pair {
-            close_pair_seeds += 1;
-        }
-    }
-
-    eprintln!("\n=== Lock-to-Airship BFS Distance ({seeds} seeds, {total_locks} locks) ===\n");
-
-    // Histogram
-    eprintln!("Distance | Count | Bar");
-    eprintln!("---------+-------+----");
-    let max_dist_with_data = histogram.iter().rposition(|&c| c > 0).unwrap_or(0);
-    for (d, &count) in histogram[..=max_dist_with_data].iter().enumerate() {
-        let bar = "#".repeat((count as usize).min(60));
-        eprintln!("{d:>5}    | {count:<5} | {bar}");
-    }
-
-    // Summary stats
-    let total_dist: u64 = histogram.iter().enumerate().map(|(d, &c)| d as u64 * c as u64).sum();
-    let mean = total_dist as f64 / total_locks.max(1) as f64;
-    let close = histogram[0] + histogram[1] + histogram[2];
-    let close_pct = close as f64 / total_locks.max(1) as f64 * 100.0;
-
-    eprintln!("\nMean distance:         {mean:.1}");
-    eprintln!("Locks at dist <= 2:    {close}/{total_locks} ({close_pct:.1}%)");
-    eprintln!("Seeds with any <= 2:   {close_lock_seeds}/{seeds}");
-    if no_target_locks > 0 {
-        eprintln!("Locks without target:  {no_target_locks}");
-    }
-
-    eprintln!("\nPer-world averages:");
-    for (wi, &(sum, count)) in per_world.iter().enumerate() {
-        if count > 0 {
-            let avg = sum as f64 / count as f64;
-            eprintln!("  W{}: {avg:.1} avg ({count} locks)", wi + 1);
-        }
-    }
-
-    eprintln!("\n=== Inter-Lock Manhattan Distance ({total_pairs} pairs) ===\n");
-    eprintln!("Distance | Count | Bar");
-    eprintln!("---------+-------+----");
-    let max_inter = inter_hist.iter().rposition(|&c| c > 0).unwrap_or(0);
-    for (d, &count) in inter_hist[..=max_inter].iter().enumerate() {
-        let bar = "#".repeat((count as usize).min(60));
-        eprintln!("{d:>5}    | {count:<5} | {bar}");
-    }
-    let inter_total: u64 = inter_hist.iter().enumerate().map(|(d, &c)| d as u64 * c as u64).sum();
-    let inter_mean = inter_total as f64 / total_pairs.max(1) as f64;
-    let close_pairs: u32 = inter_hist[..=3].iter().sum();
-    let close_pair_pct = close_pairs as f64 / total_pairs.max(1) as f64 * 100.0;
-    eprintln!("\nMean pair distance:    {inter_mean:.1}");
-    eprintln!("Pairs at dist <= 3:    {close_pairs}/{total_pairs} ({close_pair_pct:.1}%)");
-    eprintln!("Seeds with any pair <=3: {close_pair_seeds}/{seeds}");
-    eprintln!("\nPer-world pair averages:");
-    for (wi, &(sum, count)) in per_world_pairs.iter().enumerate() {
-        if count > 0 {
-            let avg = sum as f64 / count as f64;
-            eprintln!("  W{}: {avg:.1} avg ({count} pairs)", wi + 1);
-        }
-    }
-}
-
-/// Distribution analyzer for pipe placement.
-///
-/// Runs the builder for N seeds and reports, per world:
-///   - endpoint frequency (how often each position appears as a pipe end)
-///   - unordered-pair frequency
-///   - Shannon entropy of the endpoint distribution (bits)
-///   - top-5 most-picked endpoints and pairs
-///
-/// Use the entropy number to compare scoring tweaks: higher = more variety.
-/// Run with: cargo test --release test_pipe_distribution -- --ignored --nocapture
-#[test]
-#[ignore]
-fn test_pipe_distribution() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => {
-            eprintln!("ROM not found, skipping");
-            return;
-        }
-    };
-
-    let seeds: u64 = std::env::var("PIPE_SEEDS")
-        .ok()
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(1000);
-
-    // Per-world tallies
-    let mut endpoint_counts: [HashMap<(usize, usize), u32>; 8] = Default::default();
-    let mut pair_counts: [HashMap<TeleportEdge, u32>; 8] = Default::default();
-    let mut total_endpoints = [0u32; 8];
-    let mut total_pairs = [0u32; 8];
-
-    for result in &par_seeds(seeds, |seed| census_build(&rom, seed)) {
-        for built in &result.worlds {
-            let wi = built.world_idx;
-            for &(a, b) in &built.pipe_pairs {
-                *endpoint_counts[wi].entry(a).or_insert(0) += 1;
-                *endpoint_counts[wi].entry(b).or_insert(0) += 1;
-                total_endpoints[wi] += 2;
-
-                // Normalize unordered pair (smaller first)
-                let pair = if a <= b { (a, b) } else { (b, a) };
-                *pair_counts[wi].entry(pair).or_insert(0) += 1;
-                total_pairs[wi] += 1;
-            }
-        }
-    }
-
-    eprintln!("\n=== Pipe Distribution over {seeds} seeds ===");
-
-    for wi in 0..8 {
-        let expected_pairs = VANILLA_PIPE_PAIRS[wi];
-        if expected_pairs == 0 {
-            continue;
-        }
-
-        let endpoints = &endpoint_counts[wi];
-        let pairs = &pair_counts[wi];
-        let total_ep = total_endpoints[wi] as f64;
-        let total_pr = total_pairs[wi] as f64;
-
-        // Shannon entropy (bits) over endpoint distribution
-        let entropy = shannon_entropy(endpoints.values(), total_ep);
-        // Max entropy if uniform over all observed endpoints
-        let max_entropy = (endpoints.len() as f64).log2();
-
-        // Same for pairs
-        let pair_entropy = shannon_entropy(pairs.values(), total_pr);
-        let pair_max_entropy = (pairs.len() as f64).log2();
-
-        eprintln!(
-            "\n--- W{} ({} pair{}/seed) ---",
-            wi + 1,
-            expected_pairs,
-            if expected_pairs == 1 { "" } else { "s" },
-        );
-        eprintln!(
-            "  Endpoints: {} unique  |  entropy {:.2} / {:.2} bits ({:.0}%)",
-            endpoints.len(),
-            entropy,
-            max_entropy,
-            if max_entropy > 0.0 { entropy / max_entropy * 100.0 } else { 0.0 },
-        );
-        eprintln!(
-            "  Pairs:     {} unique  |  entropy {:.2} / {:.2} bits ({:.0}%)",
-            pairs.len(),
-            pair_entropy,
-            pair_max_entropy,
-            if pair_max_entropy > 0.0 { pair_entropy / pair_max_entropy * 100.0 } else { 0.0 },
-        );
-
-        let mut ep_sorted: Vec<_> = endpoints.iter().collect();
-        ep_sorted.sort_by(|a, b| b.1.cmp(a.1));
-        eprintln!("  Top endpoints:");
-        for (pos, count) in ep_sorted.iter().take(5) {
-            let count = **count;
-            let pct = count as f64 / total_ep * 100.0;
-            let bar = "#".repeat((pct as usize).min(40));
-            eprintln!(
-                "    ({:2},{:2})  {:>5} ({:5.1}%)  {bar}",
-                pos.0, pos.1, count, pct,
-            );
-        }
-
-        let mut pr_sorted: Vec<_> = pairs.iter().collect();
-        pr_sorted.sort_by(|a, b| b.1.cmp(a.1));
-        eprintln!("  Top pairs:");
-        for (pair, count) in pr_sorted.iter().take(5) {
-            let count = **count;
-            let pct = count as f64 / total_pr * 100.0;
-            let bar = "#".repeat((pct as usize).min(40));
-            eprintln!(
-                "    ({:2},{:2}) <-> ({:2},{:2})  {:>5} ({:5.1}%)  {bar}",
-                pair.0.0, pair.0.1, pair.1.0, pair.1.1, count, pct,
-            );
-        }
-    }
-
-    // One-line summary line for easy before/after diffing
-    eprintln!("\n=== Endpoint entropy summary (bits) ===");
-    let summary: Vec<String> = (0..8)
-        .filter(|&wi| VANILLA_PIPE_PAIRS[wi] > 0)
-        .map(|wi| {
-            let entropy = shannon_entropy(endpoint_counts[wi].values(), total_endpoints[wi] as f64);
-            format!("W{}={entropy:.2}", wi + 1)
-        })
-        .collect();
-    eprintln!("  {}", summary.join("  "));
-}
-
 /// Distribution analyzer for fortress placement.
 ///
 /// Runs the builder for N seeds and reports, per world:
@@ -3004,222 +1284,6 @@ fn test_fortress_distribution() {
     let expected = 17 * seeds as u32;
     eprintln!("\nGrand total: {grand_total} fortresses across {seeds} seeds (expected {expected})");
     assert_eq!(grand_total, expected, "fortress count invariant broken");
-}
-
-/// Quality analyzer for level placement.
-///
-/// Level placement is deterministic given pipes+forts, so a position-entropy
-/// test would mostly just measure upstream randomness. Instead, this measures
-/// whether the scoring achieves its stated goals:
-///
-///   - Spread: avg pairwise distance between placed levels, density-rule
-///     violations (pairs within combined radius 4). Anti-clumping is the
-///     primary anti-degeneracy signal.
-///   - Path bonus: avg detour from start→target shortest path for placed
-///     levels vs all candidates. Negative bias = levels biased toward main
-///     route, the intended design goal.
-///   - Dead-end bonus: % of dead-end candidates that became levels vs the
-///     random baseline. Treated as a tiebreaker — it should win where it
-///     doesn't conflict with path bias, but not override it.
-///
-/// Run with: cargo test --release test_level_placement_quality -- --ignored --nocapture
-/// Override seed count with LEVEL_SEEDS=N.
-#[test]
-#[ignore]
-fn test_level_placement_quality() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => {
-            eprintln!("ROM not found, skipping");
-            return;
-        }
-    };
-
-    let seeds: u64 = std::env::var("LEVEL_SEEDS")
-        .ok()
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(1000);
-
-    // Per-world aggregates
-    let mut total_pairwise_dist = [0u64; 8];
-    let mut total_pairs = [0u64; 8];
-    let mut density_violations = [0u64; 8];
-    let mut dead_ends_picked = [0u64; 8];
-    let mut total_dead_end_candidates = [0u64; 8];
-    let mut levels_picked = [0u64; 8];
-    let mut total_candidates = [0u64; 8];
-    let mut total_level_detour = [0u64; 8];
-    let mut total_levels_for_detour = [0u64; 8];
-    let mut total_candidate_detour = [0u64; 8];
-    let mut total_candidates_for_detour = [0u64; 8];
-
-    for result in &par_seeds(seeds, |seed| census_build(&rom, seed)) {
-        for built in &result.worlds {
-            let wi = built.world_idx;
-            let start_pos = rom_data::find_start(&built.grid);
-            let target_pos = find_target(&built.grid, wi);
-
-            let levels: Vec<(usize, usize)> = built.slots.iter()
-                .filter(|s| s.kind == SlotKind::Level)
-                .map(|s| s.pos)
-                .collect();
-
-            // Candidate pool seen by level placement: all non-fort, non-pipe
-            // section positions. In the final state, these became Level or
-            // HammerBro slots.
-            let candidates: Vec<(usize, usize)> = built.slots.iter()
-                .filter(|s| matches!(s.kind, SlotKind::Level | SlotKind::HammerBro))
-                .map(|s| s.pos)
-                .collect();
-
-            // BFS from start (matches what scoring used).
-            let walk = walk_map(&built.grid, &built.pipe_pairs, start_pos, built.world_idx);
-            let bfs_distances = &walk.distances;
-
-            // === Spread: avg pairwise distance between placed levels ===
-            for i in 0..levels.len() {
-                for j in (i + 1)..levels.len() {
-                    let manhattan = levels[i].0.abs_diff(levels[j].0)
-                        + levels[i].1.abs_diff(levels[j].1);
-                    total_pairwise_dist[wi] += manhattan as u64;
-                    total_pairs[wi] += 1;
-
-                    // Density rule: max(manhattan, |bfs_diff|) <= 4
-                    let bfs_diff = match (bfs_distances.get(&levels[i]), bfs_distances.get(&levels[j])) {
-                        (Some(&a), Some(&b)) => a.abs_diff(b),
-                        _ => manhattan,
-                    };
-                    if manhattan.max(bfs_diff) <= 4 {
-                        density_violations[wi] += 1;
-                    }
-                }
-            }
-
-            // === Dead-end utilization ===
-            for &pos in &candidates {
-                if is_dead_end(&built.grid, pos) {
-                    total_dead_end_candidates[wi] += 1;
-                    if levels.contains(&pos) {
-                        dead_ends_picked[wi] += 1;
-                    }
-                }
-            }
-            levels_picked[wi] += levels.len() as u64;
-            total_candidates[wi] += candidates.len() as u64;
-
-            // === Path detour: levels vs candidate baseline ===
-            // Only positions reachable in BOTH directions count — unreachable
-            // positions can't have a meaningful detour relative to a route
-            // they're not on.
-            if let Some(tp) = target_pos {
-                let reverse_walk = walk_map(&built.grid, &built.pipe_pairs, Some(tp), built.world_idx);
-                if let Some(&td) = bfs_distances.get(&tp) {
-                    for &pos in &levels {
-                        if let (Some(&fwd), Some(&rev)) = (
-                            bfs_distances.get(&pos),
-                            reverse_walk.distances.get(&pos),
-                        ) {
-                            let detour = (fwd + rev).saturating_sub(td);
-                            total_level_detour[wi] += detour as u64;
-                            total_levels_for_detour[wi] += 1;
-                        }
-                    }
-                    for &pos in &candidates {
-                        if let (Some(&fwd), Some(&rev)) = (
-                            bfs_distances.get(&pos),
-                            reverse_walk.distances.get(&pos),
-                        ) {
-                            let detour = (fwd + rev).saturating_sub(td);
-                            total_candidate_detour[wi] += detour as u64;
-                            total_candidates_for_detour[wi] += 1;
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    eprintln!("\n=== Level Placement Quality over {seeds} seeds ===");
-
-    for wi in 0..8 {
-        if total_candidates[wi] == 0 {
-            continue;
-        }
-
-        eprintln!("\n--- W{} ---", wi + 1);
-
-        // Spread
-        if total_pairs[wi] > 0 {
-            let avg_pair = total_pairwise_dist[wi] as f64 / total_pairs[wi] as f64;
-            let dens_pct = density_violations[wi] as f64 / total_pairs[wi] as f64 * 100.0;
-            eprintln!("  Spread:");
-            eprintln!("    Avg pairwise level distance: {avg_pair:.1} tiles");
-            eprintln!(
-                "    Density violations (combined radius <=4): {} / {} pairs ({:.1}%)",
-                density_violations[wi], total_pairs[wi], dens_pct,
-            );
-        }
-
-        // Dead-end bonus
-        let dead_end_util = if total_dead_end_candidates[wi] > 0 {
-            dead_ends_picked[wi] as f64 / total_dead_end_candidates[wi] as f64 * 100.0
-        } else { 0.0 };
-        let random_baseline = levels_picked[wi] as f64 / total_candidates[wi] as f64 * 100.0;
-        let lift = dead_end_util - random_baseline;
-        eprintln!("  Dead-end bonus (+0.5):");
-        eprintln!(
-            "    Dead-end candidates: {} ({:.1}% of all candidates)",
-            total_dead_end_candidates[wi],
-            total_dead_end_candidates[wi] as f64 / total_candidates[wi] as f64 * 100.0,
-        );
-        eprintln!("    Picked as level:     {dead_end_util:.1}%");
-        eprintln!("    Random baseline:     {random_baseline:.1}%");
-        eprintln!(
-            "    Lift: {lift:+.1} pp  ({})",
-            if lift.abs() < 2.0 { "negligible" }
-            else if lift > 0.0 { "bias toward dead-ends" }
-            else { "bias against dead-ends" },
-        );
-
-        // Path bonus
-        if total_levels_for_detour[wi] > 0 && total_candidates_for_detour[wi] > 0 {
-            let avg_lvl = total_level_detour[wi] as f64 / total_levels_for_detour[wi] as f64;
-            let avg_cand = total_candidate_detour[wi] as f64 / total_candidates_for_detour[wi] as f64;
-            let bias = avg_lvl - avg_cand;
-            eprintln!("  Path bonus (max = PATH_DETOUR_CAP * W_PATH):");
-            eprintln!("    Avg detour for placed levels: {avg_lvl:.2} hops");
-            eprintln!("    Avg detour for all candidates: {avg_cand:.2} hops");
-            eprintln!(
-                "    Bias: {bias:+.2} hops  ({})",
-                if bias.abs() < 0.3 { "negligible" }
-                else if bias < 0.0 { "toward main route" }
-                else { "off main route" },
-            );
-        }
-    }
-
-    // One-line summary for diffing
-    eprintln!("\n=== Summary (avg pairwise distance / dead-end lift / path bias) ===");
-    for wi in 0..8 {
-        if total_candidates[wi] == 0 { continue; }
-        let avg_pair = if total_pairs[wi] > 0 {
-            total_pairwise_dist[wi] as f64 / total_pairs[wi] as f64
-        } else { 0.0 };
-        let dead_end_util = if total_dead_end_candidates[wi] > 0 {
-            dead_ends_picked[wi] as f64 / total_dead_end_candidates[wi] as f64 * 100.0
-        } else { 0.0 };
-        let random_baseline = levels_picked[wi] as f64 / total_candidates[wi] as f64 * 100.0;
-        let lift = dead_end_util - random_baseline;
-        let path_bias = if total_levels_for_detour[wi] > 0 && total_candidates_for_detour[wi] > 0 {
-            let avg_lvl = total_level_detour[wi] as f64 / total_levels_for_detour[wi] as f64;
-            let avg_cand = total_candidate_detour[wi] as f64 / total_candidates_for_detour[wi] as f64;
-            avg_lvl - avg_cand
-        } else { 0.0 };
-        eprintln!(
-            "  W{}: dist={avg_pair:5.1}  dead-end-lift={lift:+5.1}pp  path-bias={path_bias:+5.2}",
-            wi + 1,
-        );
-    }
 }
 
 /// Regression: under start↔airship swap, the W3 fixed pipe used to be
@@ -3488,176 +1552,6 @@ fn prog_print_pass(label: &str, worlds: &[ProgLinWorld; 8]) {
     let unreach: u32 = worlds.iter().map(|w| w.unreachable).sum();
     if unreach > 0 {
         eprintln!("    WARNING: {unreach} unreachable-target case(s) — possible builder bug");
-    }
-}
-
-/// How often each of the five always-on W8 screen-3 bridges is chosen as a
-/// lock (i.e. gated "out" as a water gap until its fortress is beaten). Runs
-/// the shipped pipeline so the bridges (`apply_w8_bridges`) are present, and
-/// reports per-bridge frequency plus the distribution of how many are out per
-/// seed, for both start↔airship modes.
-///
-/// Run with: cargo test --release report_w8_bridge_locks -- --ignored --nocapture
-/// Override seed count with BRIDGE_SEEDS=N (default 2000, per mode).
-#[test]
-#[ignore]
-fn report_w8_bridge_locks() {
-    let rom_bytes = match std::fs::read("roms/Super Mario Bros. 3 (USA) (Rev 1).nes") {
-        Ok(b) => b,
-        Err(_) => {
-            eprintln!("ROM not found, skipping");
-            return;
-        }
-    };
-    let seeds: u64 = std::env::var("BRIDGE_SEEDS")
-        .ok()
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(2000);
-
-    // The 5 always-on W8 screen-3 bridges (row 5). A lock on a bridge tile
-    // gaps it out as water (`gap_tile_for`: 0xB3 -> 0x9D) until its fort falls.
-    const W8_BRIDGES: [(usize, usize); 5] = [(5, 51), (5, 53), (5, 55), (5, 57), (5, 59)];
-
-    let base = crate::Options {
-        palettes: false,
-        palette_themed: false,
-        ..Default::default()
-    };
-
-    for (label, sas) in [("SAS off", false), ("SAS on", true)] {
-        let mut opts = base.clone();
-        opts.swap_start_airship = sas;
-
-        let mut per_bridge = [0u32; 5];
-        let mut count_dist = [0u32; 6]; // index = # bridges out this seed (0..=5)
-        let mut seeds_with_any = 0u32;
-        let mut measured = 0u32;
-
-        for seed in 0..seeds {
-            let (_rom, result) =
-                match crate::randomize_rom_with_overworld_capture(&rom_bytes, seed, &opts, None) {
-                    Ok(p) => p,
-                    Err(e) => {
-                        eprintln!("seed {seed}: {e}");
-                        continue;
-                    }
-                };
-            let w8 = match result.worlds.iter().find(|b| b.world_idx == 7) {
-                Some(b) => b,
-                None => continue,
-            };
-            measured += 1;
-            let mut out_this_seed = 0;
-            for (i, &pos) in W8_BRIDGES.iter().enumerate() {
-                if w8.locks.iter().any(|l| l.pos == pos) {
-                    per_bridge[i] += 1;
-                    out_this_seed += 1;
-                }
-            }
-            count_dist[out_this_seed] += 1;
-            if out_this_seed > 0 {
-                seeds_with_any += 1;
-            }
-        }
-
-        let m = measured.max(1) as f64;
-        eprintln!("\n=== W8 bridges chosen as locks [{label}] — {measured} seeds ===");
-        eprintln!("Per-bridge (locked => bridge is out):");
-        for (i, &pos) in W8_BRIDGES.iter().enumerate() {
-            eprintln!(
-                "  bridge {} {:?}:  {:>5}  ({:>4.1}%)",
-                i + 1,
-                pos,
-                per_bridge[i],
-                per_bridge[i] as f64 / m * 100.0,
-            );
-        }
-        eprintln!("Bridges out per seed:");
-        for (n, &c) in count_dist.iter().enumerate() {
-            eprintln!("  {n}:  {:>5}  ({:>4.1}%)", c, c as f64 / m * 100.0);
-        }
-        eprintln!(
-            "Seeds with >=1 bridge out:  {seeds_with_any}/{measured}  ({:.1}%)",
-            seeds_with_any as f64 / m * 100.0,
-        );
-    }
-}
-
-/// Distribution of how many forced levels each W2 / W6 shortcut pipe skips.
-/// (Concern: these small worlds' pipes may skip too much at once.) A pipe's
-/// "skip" is the rise in minimum required clears when it's removed — i.e. the
-/// forced levels it lets the player bypass. Reports a per-pipe histogram plus
-/// mean/max for both start↔airship modes, using the shipped pipeline.
-///
-/// Run: cargo test --release report_w2_w6_pipe_skips -- --ignored --nocapture
-/// Override seed count with PROG_SEEDS=N.
-#[test]
-#[ignore]
-fn report_w2_w6_pipe_skips() {
-    let rom_bytes = match std::fs::read("roms/Super Mario Bros. 3 (USA) (Rev 1).nes") {
-        Ok(b) => b,
-        Err(_) => {
-            eprintln!("ROM not found, skipping");
-            return;
-        }
-    };
-    let seeds: u64 = std::env::var("PROG_SEEDS")
-        .ok()
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(2000);
-    let base = crate::Options {
-        palettes: false,
-        palette_themed: false,
-        ..Default::default()
-    };
-
-    for (label, sas) in [("SAS off", false), ("SAS on", true)] {
-        let mut opts = base.clone();
-        opts.swap_start_airship = sas;
-
-        // Per world_idx: histogram of Shortcut(n) skip counts (n capped at 9).
-        let mut hist: [[u32; 10]; 8] = [[0; 10]; 8];
-        for seed in 0..seeds {
-            let (_rom, result) =
-                match crate::randomize_rom_with_overworld_capture(&rom_bytes, seed, &opts, None) {
-                    Ok(p) => p,
-                    Err(_) => continue,
-                };
-            for built in &result.worlds {
-                let wi = built.world_idx;
-                for class in classify_pipes(built) {
-                    if let PipeClass::Shortcut(n) = class {
-                        hist[wi][n.min(9)] += 1;
-                    }
-                }
-            }
-        }
-
-        eprintln!("\n=== [{label}] W2 / W6 shortcut-pipe levels-skipped ({seeds} seeds) ===");
-        for &wi in &[1usize, 5] {
-            let h = &hist[wi];
-            let total: u32 = h.iter().sum();
-            let t = total.max(1) as f64;
-            let sum: u64 = h.iter().enumerate().map(|(n, &c)| n as u64 * c as u64).sum();
-            let mean = sum as f64 / t;
-            let max = h.iter().rposition(|&c| c > 0).unwrap_or(0);
-            let ge3: u32 = h[3..].iter().sum();
-            let ge4: u32 = h[4..].iter().sum();
-            eprintln!(
-                "  W{}: {total} shortcut pipes total, mean skip {mean:.2}, max {max}",
-                wi + 1
-            );
-            for (n, &c) in h.iter().enumerate() {
-                if c > 0 {
-                    eprintln!("     skips {n}: {c:>5}  ({:>4.1}%)", c as f64 / t * 100.0);
-                }
-            }
-            eprintln!(
-                "     >=3 skipped: {ge3} ({:.1}%)   >=4 skipped: {ge4} ({:.1}%)",
-                ge3 as f64 / t * 100.0,
-                ge4 as f64 / t * 100.0,
-            );
-        }
     }
 }
 
@@ -4016,131 +1910,6 @@ fn test_route_choice() {
 }
 
 
-/// Probe how a world (WORLD=n, 1-based; default 4) places connectivity pipes.
-/// For each base-arm seed it prints the world's route count plus, per pipe
-/// pair, each endpoint's walk-degree and BFS hops from start / to target. This
-/// is the tool that diagnosed the W4 junction-endpoint regression (issue #121)
-/// — junction bridges landed the island mouth next to the airship (low t),
-/// building a goal express; see `PROXIMITY_ENDPOINT_WORLDS`.
-///   PROBE_SEEDS=200 WORLD=4 cargo test --release --lib test_pipe_probe -- --ignored --nocapture
-#[test]
-#[ignore]
-fn test_pipe_probe() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => {
-            eprintln!("ROM not found, skipping");
-            return;
-        }
-    };
-    let seeds: u64 = std::env::var("PROBE_SEEDS").ok().and_then(|s| s.parse().ok()).unwrap_or(200);
-    let world: usize = std::env::var("WORLD").ok().and_then(|s| s.parse().ok()).unwrap_or(4);
-    let wi = world - 1;
-    for seed in 0..seeds {
-        if census_arm(seed) != CensusArm::Base {
-            continue; // one flag arm, so on/off compare like-for-like
-        }
-        let result = census_build(&rom, seed);
-        let Some(built) = result.worlds.iter().find(|b| b.world_idx == wi) else { continue };
-        let rc = analyze_route_choice(built, route_choice::DEFAULT_SLACK);
-        let n = if rc.reachable { rc.routes.len() } else { 0 };
-
-        let mut grid = built.grid.clone();
-        stamp_slots(&mut grid, &built.slots);
-        let start = rom_data::find_start(&grid);
-        let target = find_target(&grid, wi);
-        let d_start = walk_map(&grid, &built.pipe_pairs, start, wi).distances;
-        let d_target = target.map(|t| walk_map(&grid, &built.pipe_pairs, Some(t), wi).distances);
-        let stat = |p: (usize, usize)| -> String {
-            let deg = walk_degree(&grid, p);
-            let ds = d_start.get(&p).map(|d| d.to_string()).unwrap_or_else(|| "-".into());
-            let dt = d_target
-                .as_ref()
-                .and_then(|m| m.get(&p))
-                .map(|d| d.to_string())
-                .unwrap_or_else(|| "-".into());
-            format!("{p:?}:deg{deg},s{ds},t{dt}")
-        };
-        let pipes: Vec<String> = built
-            .pipe_pairs
-            .iter()
-            .map(|&(a, b)| format!("[{} <-> {}]", stat(a), stat(b)))
-            .collect();
-        eprintln!("seed {seed:>4}  routes {n}  cost {}  pipes {}", rc.best_cost, pipes.join(" "));
-    }
-}
-
-/// Route-choice linear% split by whether each world was Start↔Airship
-/// swapped, so the #121 junction endpoints can be checked separately under
-/// SAS on/off. Rebuilds like `census_build` but reads the per-world swap
-/// flag. Combine with PIPEMODE=proximity / FORCEJUNCTION to A/B.
-///   SAS_SEEDS=2000 cargo test --release --lib test_sas_split -- --ignored --nocapture
-#[test]
-#[ignore]
-fn test_sas_split() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => {
-            eprintln!("ROM not found, skipping");
-            return;
-        }
-    };
-    let seeds: u64 = std::env::var("SAS_SEEDS").ok().and_then(|s| s.parse().ok()).unwrap_or(2000);
-    // (world, swapped) -> (linear_count, total)
-    let per_seed = par_seeds(seeds, |seed| {
-        let arm = census_arm(seed);
-        let rom = apply_qol_variant(&rom, arm == CensusArm::HammerRocks, arm == CensusArm::EightsWild);
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let mut catalog = NodeCatalog::build(&rom, false);
-        let mut swap_rng = ChaCha8Rng::seed_from_u64(seed);
-        super::super::start_airship_swap::pick_swaps(&mut catalog, &mut swap_rng);
-        let swapped = catalog.start_airship_swapped;
-        let pickup = super::super::overworld_pickup::pick_up(
-            &rom,
-            &catalog,
-            super::super::overworld_pickup::PickupFlags {
-                shuffle_spade_games: true,
-                shuffle_toad_houses: true,
-                ..Default::default()
-            },
-        );
-        let result = build(
-            &rom,
-            &OverworldData { pickup: &pickup, catalog: &catalog },
-            &mut rng,
-            BuildFlags {
-                shuffle_toad_houses: true,
-                eights_are_wild: arm == CensusArm::EightsWild,
-                ..Default::default()
-            },
-        );
-        let mut rows = Vec::new();
-        for built in &result.worlds {
-            let rc = analyze_route_choice(built, route_choice::DEFAULT_SLACK);
-            let n = if rc.reachable { rc.routes.len() } else { 0 };
-            rows.push((built.world_idx, swapped[built.world_idx], n <= 1));
-        }
-        rows
-    });
-    // Aggregate.
-    let mut lin = [[0usize; 2]; 8];
-    let mut tot = [[0usize; 2]; 8];
-    for rows in &per_seed {
-        for &(wi, sw, is_lin) in rows {
-            let s = sw as usize;
-            tot[wi][s] += 1;
-            if is_lin { lin[wi][s] += 1; }
-        }
-    }
-    eprintln!("\n=== linear% by SAS state ({seeds} seeds) ===");
-    eprintln!("  {:<4} {:>16} {:>16}", "", "SAS-off", "SAS-on");
-    for wi in 0..8 {
-        let pct = |s: usize| if tot[wi][s] == 0 { "-".to_string() }
-            else { format!("{:.0}% (n={})", lin[wi][s] as f64 / tot[wi][s] as f64 * 100.0, tot[wi][s]) };
-        eprintln!("  W{:<3} {:>16} {:>16}", wi + 1, pct(0), pct(1));
-    }
-}
-
 /// Per-seed build wall time — the number the WASM app's generate latency
 /// tracks. Serial on purpose (per-seed timing, no thread contention).
 ///   TIME_SEEDS=40 cargo test --release --lib test_build_time -- --ignored --nocapture
@@ -4264,7 +2033,6 @@ fn test_walkgraph_reuse() {
                         replace_tile: 0x00,
                         fort_section: seed as usize % built.section_count,
                         secret_exit_safe: false,
-                        blocks_target: false,
                     });
                     assert!(
                         base.walk_invariant(&cand),
@@ -4481,139 +2249,6 @@ fn test_c1_floor_probe() {
             lin as f64 / n as f64 * 100.0,
         );
     }
-}
-
-/// Headroom probe for the gated shortcut (issue #125): per LINEAR world (the
-/// move's target), how much room is there to build one? Counts FX-slot headroom
-/// (4 locks/world cap), off-cheap-route forts (candidates to gate a shortcut
-/// with), lockless off-route forts (why a re-hunt, not an added lock), and
-/// spare-pipe presence (a shortcut needs a pipe edge).
-///
-///   ROUTE_SEEDS=500 cargo test --release --lib test_gated_shortcut_headroom -- --ignored --nocapture
-#[test]
-#[ignore]
-fn test_gated_shortcut_headroom() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let seeds: u64 = std::env::var("ROUTE_SEEDS")
-        .ok()
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(500);
-
-    // Per-seed, per-world: only linear worlds report a tuple.
-    // (fx_headroom, off_route_forts, lockless_off_route_forts, spare_pipes).
-    let per_seed = par_seeds(seeds, |seed| {
-        let result = census_build(&rom, seed);
-        let mut out: Vec<(usize, usize, usize, usize, usize)> = Vec::new();
-        for built in &result.worlds {
-            let rc = analyze_route_choice(built, route_choice::DEFAULT_SLACK);
-            if !rc.reachable || rc.routes.len() >= 2 {
-                continue; // only linear worlds are the gated-shortcut target
-            }
-            let cheap: HashSet<Pos> = rc
-                .routes
-                .first()
-                .map(|r| r.path.iter().copied().collect())
-                .unwrap_or_default();
-            let locked_sections: HashSet<usize> =
-                built.locks.iter().map(|l| l.fort_section).collect();
-            let mut off_route = 0;
-            let mut lockless_off_route = 0;
-            for s in &built.slots {
-                if s.kind != SlotKind::Fortress || cheap.contains(&s.pos) {
-                    continue;
-                }
-                off_route += 1;
-                if !locked_sections.contains(&s.section) {
-                    lockless_off_route += 1;
-                }
-            }
-            let fx_headroom = 4usize.saturating_sub(built.locks.len());
-            out.push((
-                built.world_idx,
-                fx_headroom,
-                off_route,
-                lockless_off_route,
-                built.pipe_pairs.len(),
-            ));
-        }
-        out
-    });
-
-    let mut linear_worlds = 0usize;
-    let mut with_fx_headroom = 0usize;
-    let mut with_off_route = 0usize;
-    let mut with_lockless_off_route = 0usize;
-    let mut additive_ready = 0usize; // fx headroom AND lockless off-route fort
-    let mut with_pipes = 0usize;
-    for worlds in &per_seed {
-        for &(_, fx, off, lockless, pipes) in worlds {
-            linear_worlds += 1;
-            if fx >= 1 {
-                with_fx_headroom += 1;
-            }
-            if off >= 1 {
-                with_off_route += 1;
-            }
-            if lockless >= 1 {
-                with_lockless_off_route += 1;
-            }
-            if fx >= 1 && lockless >= 1 {
-                additive_ready += 1;
-            }
-            if pipes >= 1 {
-                with_pipes += 1;
-            }
-        }
-    }
-    let pct = |n: usize| n as f64 / linear_worlds.max(1) as f64 * 100.0;
-    eprintln!("\n=== Compound-move headroom over {seeds} seeds ===");
-    eprintln!("  linear worlds (gated-shortcut target): {linear_worlds}");
-    eprintln!("  ...with FX headroom (locks<4):        {} ({:.0}%)", with_fx_headroom, pct(with_fx_headroom));
-    eprintln!("  ...with an off-cheap-route fort:      {} ({:.0}%)", with_off_route, pct(with_off_route));
-    eprintln!("  ...with a LOCKLESS off-route fort:    {} ({:.0}%)", with_lockless_off_route, pct(with_lockless_off_route));
-    eprintln!("  ...additive-ready (headroom+lockless):{} ({:.0}%)", additive_ready, pct(additive_ready));
-    eprintln!("  ...with >=1 pipe pair:                {} ({:.0}%)", with_pipes, pct(with_pipes));
-}
-
-/// Compound gated-shortcut census (issue #125): how often the joint pipe+lock
-/// move comes into play across a realistic flag mix. Reports ELIGIBLE (linear
-/// worlds that entered the search with the ingredients) and APPLIED (a gated
-/// shortcut actually placed) as rates over all built worlds.
-///
-///   ROUTE_SEEDS=500 cargo test --release --lib test_gated_shortcut_moves -- --ignored --nocapture
-#[test]
-#[ignore]
-fn test_gated_shortcut_moves() {
-    let rom = match load_rom() {
-        Some(r) => r,
-        None => return,
-    };
-    let seeds: u64 = std::env::var("ROUTE_SEEDS")
-        .ok()
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(500);
-
-    super::gated_shortcut::reset_metrics();
-    // Build every world; the gated-shortcut pass bumps its global counters in-build.
-    par_seeds(seeds, |seed| {
-        let _ = census_build(&rom, seed);
-    });
-    let eligible = super::gated_shortcut::GATED_SHORTCUT_ELIGIBLE.load(std::sync::atomic::Ordering::Relaxed);
-    let applied = super::gated_shortcut::GATED_SHORTCUT_APPLIED.load(std::sync::atomic::Ordering::Relaxed);
-    let worlds = seeds * 8;
-    eprintln!("\n=== Compound gated-shortcut over {seeds} seeds ({worlds} worlds) ===");
-    eprintln!(
-        "  eligible (linear + ingredients): {eligible} ({:.1}% of worlds)",
-        eligible as f64 / worlds as f64 * 100.0
-    );
-    eprintln!(
-        "  applied  (gated shortcut placed): {applied} ({:.1}% of worlds, {:.0}% of eligible)",
-        applied as f64 / worlds as f64 * 100.0,
-        applied as f64 / eligible.max(1) as f64 * 100.0
-    );
 }
 
 /// Production-parity route-choice render for ONE real seed: runs the full
